% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_blooddata.R
\name{bd_addfit}
\alias{bd_addfit}
\title{Add a model fit to a blooddata object}
\usage{
bd_addfit(
  blooddata,
  fit,
  modeltype = c("Blood", "BPR", "parentFraction", "AIF")
)
}
\arguments{
\item{blooddata}{A blooddata object created using one of the
create_blooddata_* functions.}

\item{fit}{A model fit object, for which the predict function can be used for
new data. The fit must predict from time alone, because that is how
\code{\link{bd_getdata}} calls it. Fits that model many measurements
jointly and so need a grouping variable, such as
\code{\link{metab_hgam}}, cannot be stored this way; predict for the
measurement in question and use \code{\link{bd_addfitted}} instead.}

\item{modeltype}{The function which the model predicts. One of the following:
Blood for models of how the blood data should be described, BPR for models
of the blood-to-plasma ratio, parentFraction for models of metabolism, and
AIF for models of the arterial input function.}
}
\value{
A blooddata object with the model inserted.
}
\description{
This function adds a fit object which can predict values to the blooddata
object
}
\examples{
\dontrun{
blooddata <- bd_addfit(blooddata, hillfit, "parentFraction")
}
}
\seealso{
\code{\link{bd_addfitted}} for models whose predictions must be
  supplied as values, and \code{\link{bd_addfitpars}} for models fully
  described by their parameters.
}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
