% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_blooddata.R
\name{bd_blood_dispcor}
\alias{bd_blood_dispcor}
\title{Perform Dispersion Correction on a blooddata object}
\usage{
bd_blood_dispcor(blooddata, tau, timedelta = NULL,
  keep_interpolated = T, smooth_iterations = 0)
}
\arguments{
\item{blooddata}{A blooddata object created using one of the
create_blooddata_* functions.}

\item{tau}{Time constant denoting the time of dispersion (in seconds).}

\item{timedelta}{The time difference between each measured sample. Defaults
to the most common time difference between the first 20 measurements.}

\item{keep_interpolated}{Defaults to TRUE: should interpolated samples which
were not included in the original input be included in the output.}

\item{smooth_iterations}{The number of times that the smoothing of each pair
of observations should be performed using blood_smooth. Defaults to 0.}
}
\value{
A blooddata object after dispersion correction.
}
\description{
Dispersion corrected is performed on the continuous blood data if it has not
already been performed.
}
\examples{
\dontrun{
blooddata <- bd_blood_dispcor(blooddata, 2.5)
}

}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
