% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_blooddata.R
\name{bd_extract}
\alias{bd_extract}
\title{Extract data from blooddata objects}
\usage{
bd_extract(
  blooddata,
  output = c("Blood", "BPR", "parentFraction", "AIF"),
  startTime = 0,
  stopTime = NULL,
  interpPoints = 6000,
  what = c("raw", "pred", "interp")
)
}
\arguments{
\item{blooddata}{The blooddata object.}

\item{output}{What curve should be extracted. It can be "Blood", "BPR"
(whole-blood-to-plasma-ratio), "parentFraction" or "AIF".}

\item{startTime}{The starting time for the interpolation. Defaults to zero.
If, after application of the TimeShift value in the blooddata object, the
startTime is still after zero, it will be set to zero.}

\item{stopTime}{The end time for the interpolation. Defaults to the maximum
measured time.}

\item{interpPoints}{The number of points to interpolate over between the
start and stop times. Defaults to 6000.}

\item{what}{What kind of data should be extracted? The "raw" data is the
unaltered data for modelling, the "pred" data is the predicted data at the
times of the original samples for assessing correspondence, and the
"interp" values are the values interpolated through the interptimes.}
}
\value{
A tibble with the relevant data
}
\description{
Pull out data for modelling purposes.
}
\examples{
\dontrun{
bd_extract(blooddata, output="parentFraction", what="raw")
}
}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
