% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_blooddata.R
\name{bd_getdata}
\alias{bd_getdata}
\title{Create an input object from a blooddata object}
\usage{
bd_getdata(
  blooddata,
  startTime = 0,
  stopTime = NULL,
  interpPoints = 6000,
  output = "input",
  bpr_peakfrac_cutoff = 0.001
)
}
\arguments{
\item{blooddata}{A blooddata object.}

\item{startTime}{The starting time for the interpolation.}

\item{stopTime}{The end time for the interpolation.}

\item{interpPoints}{The number of points to interpolate over.}

\item{output}{The output.}

\item{bpr_peakfrac_cutoff}{The cutoff for the blood-to-plasma ratio.}
}
\value{
Nothing. Called for its error.
}
\description{
DEFUNCT: this function read the old PET BIDS blood structure, and has been
replaced by \code{bd_extract()} for individual curves and
\code{bd_create_input()} for a whole input object. Calling it is now an
error. Blooddata objects saved in the old structure can be brought up to
date using \code{update_blooddata()}, or better, created afresh from the
source.
}
\examples{
\dontrun{
# Instead of bd_getdata(blooddata), use
bd_create_input(blooddata)

# and instead of bd_getdata(blooddata, output = "parentFraction"), use
bd_extract(blooddata, output = "parentFraction")
}
}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
