% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_blooddata.R
\name{bd_tidy_times}
\alias{bd_tidy_times}
\title{Tidy time information for blooddata operations}
\usage{
bd_tidy_times(blooddata, startTime = 0, stopTime = NULL, interpPoints = 6000)
}
\arguments{
\item{blooddata}{A blooddata object.}

\item{startTime}{The starting time for the interpolation. Defaults to zero. If, after application of the TimeShift value in the blooddata object, the startTime is still after zero, it will be set to zero.}

\item{stopTime}{The end time for the interpolation. Defaults to the maximum measured time.}

\item{interpPoints}{The number of points to interpolate over between the start and stop times. Defaults to 6000.}
}
\value{
The startTime, stopTime and interptimes after the relevant checks
}
\description{
Just a little tidier function.
}
\examples{
\dontrun{
bd_tidy_times(blooddata)
}
}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
