% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_bids.R
\name{bids_create_blooddata}
\alias{bids_create_blooddata}
\title{Create a blooddata object from BIDS data}
\usage{
bids_create_blooddata(filedata)
}
\arguments{
\item{filedata}{A table of the file data, created using bids_parse_files()}
}
\value{
A blooddata object
}
\description{
This function takes a set of files from one measurement of a BIDS-structured
study, reads in the data, and creates a blooddata file.
}
\examples{
\dontrun{
bd <- bids_create_blooddata(filedata)
}
}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
