% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_bids.R
\name{bids_filename_attributes}
\alias{bids_filename_attributes}
\title{Extract BIDS attributes from filenames and file paths}
\usage{
bids_filename_attributes(filename)
}
\arguments{
\item{filename}{The filename and filepath of the relevant file. Recommended
to provide the relative path within the studypath.}
}
\value{
A tibble containing all the data
}
\description{
Get all the stuff into a table
}
\examples{
bids_filename_attributes(
   "sub-01/ses-01/pet/sub-01_ses-01_recording-continuous_blood.json")
}
