% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_bids.R
\name{bids_parse_files}
\alias{bids_parse_files}
\title{Extract the filenames from a PET BIDS study}
\usage{
bids_parse_files(studypath)
}
\arguments{
\item{studypath}{The BIDS study path for the main study.}
}
\value{
A data frame of the files, what they are, and where they are
}
\description{
This function returns a data frame of the files for each measurement in a
PET BIDS study, its path, and what it is.
}
\examples{
\dontrun{
filedata <- bids_parse_files(studypath)
}
}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
