% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_bids.R
\name{bids_parse_study}
\alias{bids_parse_study}
\title{Parse the contents of a PET BIDS study}
\usage{
bids_parse_study(studypath)
}
\arguments{
\item{studypath}{The BIDS study path for the main study.}
}
\value{
A data frame of the files, what they are, and where they are
}
\description{
This function parses a PET BIDS study, and returns a nested tibble with the
relevant information nested appropriately.
}
\examples{
\dontrun{
studydata <- bids_parse_study(studypath)
}
}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
