% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_bloodmodels.R
\name{blmod_exp_sep}
\alias{blmod_exp_sep}
\title{Fit an exponential model to AIF data using the measured peak}
\usage{
blmod_exp_sep(time, activity, Method = NULL, weights = NULL,
  fit_exp3 = TRUE, peaktime_val = NULL, peakval = NULL, t0 = NULL,
  lower = NULL, upper = NULL, start = NULL, multstart_lower = NULL,
  multstart_upper = NULL, multstart_iter = 100, taper_weights = TRUE,
  check_startpars = FALSE)
}
\arguments{
\item{time}{The time of each measurement in seconds}

\item{activity}{The radioactivity of each measurement}

\item{Method}{Optional. The method of collection, i.e. "Discrete" or
"Continuous"}

\item{weights}{Optional. Weights of each measurement.}

\item{fit_exp3}{Should the third exponential be fitted, or should a
bi-exponential model be used? Default is TRUE for a tri-exponential model.}

\item{peaktime_val}{Optional. This allows the user to specify the peaktime
value which will be used. Otherwise it will be specified from the data.}

\item{peakval}{Optional. This allows the user to specify the peak value
which will be used. Otherwise it will be specified from the data.}

\item{t0}{Optional. This allows the user to specify the t0 value
which will be used. Otherwise it will be estimated from the data using a
linear regression.}

\item{lower}{Optional. The lower limits of the fit. If left as NULL, they
will be given reasonable defaults (mostly 50\% of the starting parameters).}

\item{upper}{Optional. The upper limits of the fit. If left as NULL, they
will be given reasonable defaults (mostly 150\% of the starting
parameters).}

\item{start}{Optional. The starting parameters for the fit. If left as NULL,
they will be selected using \code{blmod_exp_startpars}.}

\item{multstart_lower}{Optional. The lower limits of the starting parameters.}

\item{multstart_upper}{Optional. The upper limits of the starting parameters.}

\item{multstart_iter}{The number of fits to perform with different starting
parameters. If set to 1, then the starting parameters will be used for a
single fit.}

\item{taper_weights}{Should the weights be tapered to gradually trade off
between the continuous and discrete samples after the peak?}

\item{check_startpars}{Optional. Return only the starting parameters. Useful
for debugging fits which do not work.}
}
\value{
A model fit including all of the individual parameters, fit details,
  and model fit object of class blood_exp.
}
\description{
This model fits a bi- or tri-exponential model to AIF data. This model, in
contrast to the \code{blmod_expsep()} model, uses the measured peak as
the peak, and describes the rise by interpolating between t0 and the peak,
and fits an exponential model only to the fall of the curve. This model is
more robust, but less flexible than the fitpeak model.
}
\details{
The model can fit two or three exponentials to the curve
after the peak (\code{fit_exp0}). For \code{t0}, \code{peaktime} and
\code{peakval}, the values can be specified by the user, otherwise they are
determined from the data. \code{t0} is determined by fitting a regression
line through the rise, while \code{peaktime} and \code{peakval} are just
the x and y values of the maximal point in the measured data.
}
\examples{
blooddata <- create_blooddata_bids(pbr28$jsondata[[3]])
blooddata <- bd_blood_dispcor(blooddata,
                 tau = blooddata$Data$Blood$Continuous$DispersionConstant)
aif <- bd_getdata(blooddata, output = "AIF")
blood_fit <- blmod_exp_sep(aif$time,
                           aif$aif,
                           Method = aif$Method, multstart_iter = 1)
}
