% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_bloodmodels.R
\name{blmod_exp_startpars}
\alias{blmod_exp_startpars}
\title{Create starting parameters for an exponential blood model}
\usage{
blmod_exp_startpars(
  time,
  activity,
  fit_exp3 = TRUE,
  expdecay_props = c(1/60, 0.1)
)
}
\arguments{
\item{time}{The time of each measurement.}

\item{activity}{The radioactivity of each measurement.}

\item{fit_exp3}{Should a third exponential be fitted, or is a bi-exponential
fit desired? Default is TRUE.}

\item{expdecay_props}{What proportions of the decay should be used for
choosing starting parameters for the exponential decay. Defaults to 1/60
and 1/10, i.e. start to 1/60, 1/60 to 1/10 and 1/10 to end. If fitting only
two exponentials, the second value will be used.}
}
\value{
A list of the starting parameters
}
\description{
This function guesses reasonable starting parameters for exponential blood
models. I use the suggestions from page 515 of Pinheiro & Bates Mixed-Effects
Models in S and S-PLUS for creating the exponential starting parameters.
}
\examples{
blooddata <- create_blooddata_bids(pbr28$jsondata[[1]])
blooddata <- bd_blood_dispcor(blooddata)
aif <- bd_extract(blooddata, output = "AIF")
start <- blmod_exp_startpars(aif$time,
                           aif$aif)
}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
