% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_bloodmodels.R
\name{blmod_feng_model}
\alias{blmod_feng_model}
\title{Feng model for fitting of arterial input functions}
\usage{
blmod_feng_model(time, t0, A, alpha, B, beta, C, gamma)
}
\arguments{
\item{time}{Time of each sample.}

\item{t0}{The delay time. This is the point at which the linear rise begins.}

\item{A}{The multiplier of the first exponential.}

\item{alpha}{The rate of the first exponential.}

\item{B}{The multiplier of the second exponential.}

\item{beta}{The rate of the second exponential.}

\item{C}{The multiplier of the third exponential.}

\item{gamma}{The rate of the third exponential.}
}
\value{
Model predictions
}
\description{
This is the model itself for the Feng model of the AIF.
}
\examples{
blmod_feng_model(1:1000, 30, 7, 0.02, 2, 0.005, 4.5, 0.0005)
}
