% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_bloodmodels.R
\name{blmod_feng_startpars}
\alias{blmod_feng_startpars}
\title{Create starting parameters for an Feng blood model}
\usage{
blmod_feng_startpars(time, activity, expdecay_props = c(0, 0.5))
}
\arguments{
\item{time}{The time of each measurement.}

\item{activity}{The radioactivity of each measurement.}

\item{expdecay_props}{What proportions of the decay should be used for
choosing starting parameters for the exponential decay. Defaults to 0
and 0.5, i.e. the latter two exponentials, B, C, beta, gamma are estimated
using halfway to the end of the decay, and the beginning to halfway through
the decay. The first parameters, A and alpha, are estimated from the ascent.}
}
\value{
A list of the starting parameters
}
\description{
This function estimates reasonable starting parameters for the Feng blood
models. I use the suggestions from Tonietto (2012).
}
\examples{
blooddata <- pbr28$blooddata[[1]]
blooddata <- bd_blood_dispcor(blooddata)
aif <- bd_extract(blooddata, output = "AIF")
start <- blmod_feng_startpars(aif$time,
                           aif$aif)
}
\references{
Tonietto, M. (2012). Methods and models for the characterization
of arterial input function in dynamic PET studies. Master thesis.
}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
