% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_bloodmodels.R
\name{blmod_fengconv_model}
\alias{blmod_fengconv_model}
\title{Convolved Feng model for fitting of arterial input functions}
\usage{
blmod_fengconv_model(time, t0, A, alpha, B, beta, C, gamma, ti)
}
\arguments{
\item{time}{Time of each sample.}

\item{t0}{The delay time. This is the point at which the linear rise begins.}

\item{A}{The multiplier of the first exponential.}

\item{alpha}{The rate of the first exponential.}

\item{B}{The multiplier of the second exponential.}

\item{beta}{The rate of the second exponential.}

\item{C}{The multiplier of the third exponential.}

\item{gamma}{The rate of the third exponential.}

\item{ti}{The infusion time of the bolus.}
}
\value{
Model predictions
}
\description{
This is the model itself for the Fengconv model of the AIF with an extended
injection.
}
\examples{
blmod_fengconv_model(1:1000, 30, 220, 0.4, 100, 0.05, 22, 0.003, 30)
}
