% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_blooddata.R
\name{blood_addfit}
\alias{blood_addfit}
\title{Add a model fit to a blooddata object}
\usage{
blood_addfit(blooddata, fit, modeltype = c("Blood", "BPR",
  "parentFraction", "AIF"))
}
\arguments{
\item{blooddata}{A blooddata object created using one of the
create_blooddata_* functions.}

\item{fit}{A model fit object, for which the predict function can be used for
new data.}

\item{modeltype}{The function which the model predicts. One of the following:
Blood for models of how the blood data should be described, BPR for models
of the blood-to-plasma ratio, parentFraction for models of metabolism, and
AIF for models of the arterial input function.}
}
\value{
A blooddata object with the model inserted.
}
\description{
This function adds a fit object which can predict values to the blooddata
object
}
\examples{
\dontrun{
blooddata <- blood_addfit(blooddata, hillfit, "parentFraction")
}
}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
