% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_blooddata.R
\name{blood_addfitted}
\alias{blood_addfitted}
\title{Add fitted values from another fit to a blooddata object}
\usage{
blood_addfitted(blooddata, time, predicted, modeltype = c("Blood", "BPR",
  "parentFraction", "AIF"))
}
\arguments{
\item{blooddata}{A blooddata object created using one of the create_blooddata_* functions.}

\item{time}{The times of the predictions at a sufficiently detailed level of interpolation.}

\item{predicted}{The predicted values for the given set of times.}

\item{modeltype}{The function which the model predicts. One of the following:
Blood for models of how the blood data should be described, BPR for models
of the blood-to-plasma ratio, parentFraction for models of metabolism, and
AIF for models of the arterial input function.}
}
\value{
A blooddata object with the model fits inserted.
}
\description{
This function adds the predicted values of an external function to a blooddata object. This allows one to use functions outside of kinfitr to estimate blood values.
}
\examples{
\dontrun{
blooddata <- blood_addfitpars(blooddata, matlabout$time,
             matlabout$predicted, "AIF")
}
}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
