% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_bloodfuncs.R
\name{blood_dispcor}
\alias{blood_dispcor}
\title{Perform dispersion correction on blood data collected using ABSS}
\usage{
blood_dispcor(
  time,
  activity,
  tau,
  timedelta = NULL,
  keep_interpolated = T,
  smooth_iterations = 0
)
}
\arguments{
\item{time}{Vector of time (in seconds) of each sample.}

\item{activity}{Vector of measured radioactivity concentrations in  blood
samples.}

\item{tau}{Time constant denoting the time of dispersion (in seconds).}

\item{timedelta}{The time difference between each measured sample. Defaults
to the most common time difference between the first 20 measurements.}

\item{keep_interpolated}{Defaults to TRUE: should interpolated samples which
were not included in the original input be included in the output.}

\item{smooth_iterations}{The number of times that the smoothing of each pair
of observations should be performed using blood_smooth. Defaults to 0.}
}
\value{
A tibble containing the measured times and radioactivity
  concentrations after dispersion correction.
}
\description{
Perform dispersion correction on continuous blood samples from vectors as per
http://www.turkupetcentre.net/petanalysis/input_dispersion.html. The measured
values are first interpolated so that all measurements have an equal time
between them, so any missing measurements are interpolated.
}
\examples{
time <- 1:20
activity <- rnorm(20)
tau <- 2.5

blood_dispcor(time, activity, tau, 1)
}
