% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_bloodfuncs.R
\name{blood_smooth}
\alias{blood_smooth}
\title{Smooth continuous blood data}
\usage{
blood_smooth(time, activity, iterations = 1)
}
\arguments{
\item{time}{Vector of time (in seconds) of each sample.}

\item{activity}{Vector of measured radioactivity concentrations in  blood
samples.}

\item{iterations}{The number of times that the smoothing of each pair of
observations should be performed.}
}
\value{
A tibble containing the measured times and radioactivity
  concentrations after smoothing.
}
\description{
This function averages each pair of blood measurements from ABSS systems -
can be used after dispersion correction.
}
\examples{
time <- 1:20
activity <- rnorm(20)

blood_smooth(time, activity, 5)
}
