% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_bloodstream.R
\name{bloodstream_import_aifpars}
\alias{bloodstream_import_aifpars}
\title{Import bloodstream AIF fitted parameters}
\usage{
bloodstream_import_aifpars(blstream_folder)
}
\arguments{
\item{blstream_folder}{The specific bloodstream derivative folder (for the whole study) from which the data should be extracted}
}
\value{
The BIDS attributes and fitted AIF parameters.
}
\description{
Imports bloodstream AIF fitted parameters for a BIDS dataset, and converts all the units to minutes and kBq.
}
\examples{
\dontrun{
bloodstream_import_aifpars("../OpenNeuro/ds004230/derivatives/bloodstream2022-11-24_id-9AAO/")
}
}
