% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_bloodstream.R
\name{bloodstream_import_inputfunctions}
\alias{bloodstream_import_inputfunctions}
\title{Import bloodstream input functions}
\usage{
bloodstream_import_inputfunctions(blstream_folder)
}
\arguments{
\item{blstream_folder}{The specific bloodstream derivative folder (for the whole study) from which the data should be extracted}
}
\value{
The BIDS attributes and kinfitr input objects
}
\description{
Imports input function data from bloodstream for a BIDS dataset, and converts all the units to minutes and kBq.
}
\examples{
\dontrun{
bloodstream_import_inputfunction("../OpenNeuro/ds004230/derivatives/bloodstream2022-11-24_id-9AAO/")
}
}
