% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_blooddata.R
\name{create_blooddata_bids}
\alias{create_blooddata_bids}
\title{Create a blooddata object from BIDS data}
\usage{
create_blooddata_bids(bids_data)
}
\arguments{
\item{bids_data}{The filename of a PET BIDS json sidecar, or a list
containing the information contained within a PET BIDS json sidecar.}
}
\value{
A blooddata object
}
\description{
This function creates a blooddata object from data structured according to
the PET BIDS standard.
}
\examples{
\dontrun{
a <- create_blooddata_bids('bids_sidecar.json')
}
}
