% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_miscfuncs.R
\name{decay_correct}
\alias{decay_correct}
\title{Perform Decay Correction of TAC values}
\usage{
decay_correct(t_start, t_end, tac_uncor, radioisotope = c("C11", "O15",
  "F18"))
}
\arguments{
\item{t_start}{The starting times of the frames in minutes.}

\item{t_end}{The end times of the frames in minutes.}

\item{tac_uncor}{The uncorrected radioactivity values.}

\item{radioisotope}{The radioisotope.}
}
\value{
The tac after decay correction
}
\description{
This function uses the equation described here to perform the calculations: http://www.turkupetcentre.net/petanalysis/decay.html
}
\examples{
data(pbr28)
s1 <- pbr28$tacs[[1]]

# Assuming the data were not decay-corrected (they are)
decay_correct(
  s1$StartTime/60,
 (s1$StartTime + s1$Duration)/60,
 tac_uncor = s1$FC)

}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
