% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_dock.R
\name{dock_find_tracer}
\alias{dock_find_tracer}
\title{Find tracers in the DOCK database}
\usage{
dock_find_tracer(tracer_grob)
}
\arguments{
\item{tracer_grob}{A character string to search for among tracer names
(case-insensitive). For example, \code{"raclopride"}, \code{"PBR"},
or \code{"WAY"}.}
}
\value{
A tibble with columns \code{tracer} and \code{target}, containing
  all matching tracers. Returned invisibly.
}
\description{
Searches the DOCK (Database of CNS PET Kinetic parameters) database for
tracers matching a search string. Requires an internet connection to query
the database from GitHub.
}
\details{
The DOCK database is described in Kudo et al. (2024). Database of
  CNS PET kinetic parameters (DOCK): a new database of average
  parameter estimates for neuroreceptors and transporters. Annals of
  Nuclear Medicine. DOI: 10.1007/s12149-024-01947-z
}
\examples{
\dontrun{
dock_find_tracer("raclopride")
dock_find_tracer("PBR")
dock_find_tracer("WAY")
}

}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
