% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_nested_utils.R
\name{.nested_num}
\alias{.nested_num}
\title{Format a fixed parameter value for use in a delta method expression}
\usage{
.nested_num(x)
}
\arguments{
\item{x}{Numeric value.}
}
\value{
A character string of the value, parenthesised.
}
\description{
Format a fixed parameter value for use in a delta method expression
}
\keyword{internal}
