% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_feng_1tc_ref.R
\name{feng_1tc_tac}
\alias{feng_1tc_tac}
\title{Feng input model convolved with a 1TC IRF}
\usage{
feng_1tc_tac(
  t_tac,
  tac,
  weights = NULL,
  fit_t0 = TRUE,
  frameStartEnd = NULL,
  timeStartEnd = NULL,
  multstart_iter = 500
)
}
\arguments{
\item{t_tac}{Numeric vector of times for each frame in minutes. We use the
time halfway through the frame as well as a zero. If a time zero frame is
not included, it will be added.}

\item{tac}{Numeric vector of radioactivity concentrations in the tissue for
each frame. We include zero at time zero: if not included, it is added.}

\item{weights}{Optional. Numeric vector of the weights assigned to each frame
in the fitting. We include zero at time zero: if not included, it is added.
If not specified, uniform weights will be used.}

\item{fit_t0}{Should a time zero point be fitted? If TRUE, the model can
accommodate zero values in the TAC before rising. If FALSE, the TAC must
rise at time = 0.}

\item{frameStartEnd}{Optional. This allows one to specify the beginning and
final frame to use for modelling, e.g. c(1,20). This can be used to assess
time stability for example.}

\item{timeStartEnd}{Optional. This allows one to specify the beginning and
end time point instead of defining the frame numbers using frameStartEnd.
This function will restrict the model to all time frames whose t_tac is
between the values, i.e. c(0,5) will select all frames with midtimes
during the first 5 minutes.}

\item{multstart_iter}{Number of iterations for starting parameters. Default
is 500. Given as a single number, starting values are chosen by improved
Latin hypercube sampling rather than at random; given as one value per
parameter, they form a Cartesian grid. For more information, see
\code{\link[nls.multstart]{nls_multstart}}.}
}
\value{
A list with a data frame of the fitted parameters \code{out$par},
  their percentage standard errors \code{out$par.se}, the model fit object
  \code{out$fit}, the model weights \code{out$weights}, and a dataframe
  containing the TACs both of the data and the fitted values \code{out$tacs}.
}
\description{
This function is intended for describing a TAC, such as the reference region
TAC. The parameters are not meant to be interpreted as true values, but
merely as a guide for interpolating the measured curve through a parametric
description. For this reason, there are no upper and lower limits, because
completely incorrect values can be helpful for interpolating our measured
TACs more closely.
}
\examples{

data(simref)

t_tac <- simref$tacs[[2]]$Times
tac <- simref$tacs[[2]]$Reference
weights <- simref$tacs[[2]]$Weights

fit <- feng_1tc_tac(t_tac, tac, weights)
}
\references{
Jiao, J. et al, 2023. NiftyPAD-Novel Python Package for
  Quantitative Analysis of Dynamic PET Data. Neuroinformatics, pp.1-12.
  Matheson, G.J & Ogden, R.T., in preparation. SiMBA for Reference Tissue
  Models.
}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
