% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_feng_1tc_ref.R
\name{feng_1tc_tac_model}
\alias{feng_1tc_tac_model}
\title{Model: Feng input model convolved with a 1TC IRF}
\usage{
feng_1tc_tac_model(t_tac, t0, A, B, C, alpha, beta, gamma, Ph1, Th1)
}
\arguments{
\item{t_tac}{Numeric vector of time values in minutes.}

\item{t0}{The time point at which the curve begins to increase}

\item{A}{Feng AIF model A parameter}

\item{B}{Feng AIF model B parameter}

\item{C}{Feng AIF model C parameter}

\item{alpha}{Feng AIF model alpha parameter}

\item{beta}{Feng AIF model beta parameter}

\item{gamma}{Feng AIF model gamma parameter}

\item{Ph1}{1TC IRF model K1 parameter}

\item{Th1}{1TC IRF model k2 parameter}
}
\value{
The predicted values
}
\description{
This is the model definition for the Feng input model convolved with a 1TC
IRF.
}
\examples{
data(simref)
t_tac <- simref$tacs[[2]]$Times

feng_1tc_tac_model(t_tac, 0, 3, 1, 0.2, 0.6, 0.2, 0.01, 0.2, 0.1)
}
\references{
Jiao, J. et al, 2023. NiftyPAD-Novel Python Package for
  Quantitative Analysis of Dynamic PET Data. Neuroinformatics, pp.1-12.
  Matheson, G.J & Ogden, R.T., in preparation. SiMBA for Reference Tissue
  Models.
}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
