% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_miscfuncs.R
\name{frame_cumsum}
\alias{frame_cumsum}
\title{Cumulative integral over frames}
\usage{
frame_cumsum(dur, tac)
}
\arguments{
\item{dur}{Frame durations in minutes.}

\item{tac}{Radioactivity concentrations for each frame.}
}
\value{
The cumulative integral of radioactivity.
}
\description{
This function uses the frame radioactivity values and durations to create a
cumulative integral. It sums the multiple of radioactivity and duration
for all frames preceding the current frame, and halves that of the current
frame. This is an alternative to using the trapezoidal method.
}
\examples{
data(pbr28)

tac <- pbr28$tacs[[2]]$WB
dur <- pbr28$tacs[[2]]$Duration / 60

frame_integ <- frame_cumsum(dur, tac)


}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
