% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_frtm.R
\name{frtm_model}
\alias{frtm_model}
\title{Model: Full Reference Tissue Model}
\usage{
frtm_model(t_tac, reftac, R1, k2, bp, k4)
}
\arguments{
\item{t_tac}{Numeric vector of times for each frame in minutes. We use the time halfway through the frame as well as a zero.}

\item{reftac}{Numeric vector of radioactivity concentrations in the reference tissue for each frame.}

\item{R1}{Parameter value for R1}

\item{k2}{Parameter value for k2}

\item{bp}{Parameter value for bp}

\item{k4}{Parameter value for k4}
}
\value{
A numeric vector of the predicted values of the TAC in the target region.
}
\description{
This is the FRTM model itself by which predicted values are generated.
}
\examples{
\dontrun{
frtm_model(t_tac, reftac, R1 = 0.9, k2 = 0.1, bp = 1.5, k4 = 0.1)
}

}
\references{
Cunningham, Vincent J., et al. "Compartmental analysis of
diprenorphine binding to opiate receptors in the rat in vivo and its
comparison with equilibrium data in vitro." Journal of Cerebral Blood Flow
& Metabolism 11.1 (1991): 1-9.
}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
