% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_miscfuncs.R
\name{get_se}
\alias{get_se}
\title{Estimate the standard error of parameters using the delta method}
\usage{
get_se(fit, expression)
}
\arguments{
\item{fit}{A fit object for a particular model - in this case the actual lm
or nls object}

\item{expression}{The expression for which the SE is being calculated}
}
\value{
The SE as a proportion of the parameter estimate.
}
\description{
This function returns the SE of parameters, as a proportion of their
parameter estimates, calculated using the delta method as implemented in the
car package. If the parameter does not exist, this function will return NA.
}
\examples{
data(mtcars)
a <- lm(mpg ~ cyl + disp, data=mtcars)

get_se(a, "cyl")
get_se(a, "cyl/disp")

}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
