% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_telemetry.R
\name{get_telemetry}
\alias{get_telemetry}
\title{Get Telemetry Data}
\usage{
get_telemetry(
  url = "http://54.144.240.214/check/kinfitr/",
  number_of_records = 0
)
}
\arguments{
\item{url}{The url endpoint}

\item{number_of_records}{How many records. 0 for all.}
}
\value{
A list of the recorded telemetry data
}
\description{
Checks to see what has been posted to the URL endpoint. Should return
location and any any other data that gets put there with send_telemetry
}
