% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_miscfuncs.R
\name{kinfit_convolve}
\alias{kinfit_convolve}
\title{Convolution Function}
\usage{
kinfit_convolve(a, b, stepsize)
}
\arguments{
\item{a}{A numeric vector of TAC function at equally spaced intervals}

\item{b}{A numeric vector of TAC function at equally spaced intervals of equal length to \code{a}}

\item{stepsize}{The time interval between each value of \code{a} and \code{b}}
}
\value{
A numeric vector of the convolved function of the same length as \code{a} and \code{b}
}
\description{
Function to convolve two vectors of equal length for kinetic modelling.
}
\examples{
a <- rnorm(20)
b <- rnorm(20)
stepsize = 1

kinfit_convolve(a,b,stepsize)

}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
