% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_lin2tcm.R
\name{lin2tcm}
\alias{lin2tcm}
\title{Linear Two Tissue Compartment Model}
\usage{
lin2tcm(
  t_tac,
  tac,
  input,
  weights = NULL,
  inpshift = 0,
  vB = NULL,
  frameStartEnd = NULL
)
}
\arguments{
\item{t_tac}{Numeric vector of times for each frame in minutes. We use the time halfway through the frame as well as a
zero. If a time zero frame is not included, it will be added.}

\item{tac}{Numeric vector of radioactivity concentrations in the target tissue for each frame. We include zero at time
zero: if not included, it is added.}

\item{input}{Data frame containing the blood, plasma, and parent fraction concentrations over time.  This can be generated
using the \code{blood_interp} function.}

\item{weights}{Optional. Numeric vector of the weights assigned to each frame in the fitting. We include zero at time zero:
if not included, it is added. If not specified, uniform weights will be used.}

\item{inpshift}{Optional. The number of minutes by which to shift the timing of the input data frame forwards or backwards.
If not specified, this will be set to 0. This can be fitted using 1TCM or 2TCM.}

\item{vB}{Optional. The blood volume fraction.  If not specified, this will
be fitted. If specified as a numer (e.g. 0.05 for 5%), then that value will
be used.}

\item{frameStartEnd}{Optional: This allows one to specify the beginning and final frame to use for modelling, e.g. c(1,20).
This is to assess time stability.}
}
\value{
A list with a data frame of the fitted parameters \code{out$par}, the model fit object \code{out$fit},
a dataframe containing the TACs of the data \code{out$tacs}, a dataframe containing the fitted values \code{out$fitvals},
the blood input data frame after time shifting \code{input}, a vector of the weights \code{out$weights},
the inpshift value used \code{inpshift} and the specified vB value \code{out$vB}.
}
\description{
Function to fit the linearised 2TCM function.
}
\examples{
data(pbr28)

t_tac <- pbr28$tacs[[2]]$Times / 60
tac <- pbr28$tacs[[2]]$FC
weights <- pbr28$tacs[[2]]$Weights

input <- blood_interp(
  pbr28$procblood[[2]]$Time / 60, pbr28$procblood[[2]]$Cbl_dispcorr,
  pbr28$procblood[[2]]$Time / 60, pbr28$procblood[[2]]$Cpl_metabcorr,
  t_parentfrac = 1, parentfrac = 1
)

fit1 <- lin2tcm(t_tac, tac, input, weights, inpshift=0.1)
fit2 <- lin2tcm(t_tac, tac, input, weights, inpshift = 0.1, vB=0.05)
}
\references{
Oikonen, V (2003). Multilinear solution for 4-compartment model:
I. Tissue compartments in series.
Gjedde A, Wong DF 1990. Modeling neuroreceptor
binding of radioligands in vivo. In: Quantitative imaging: neuroreceptors,
neurotransmitters, and enzymes. (Eds. Frost JJ, Wagner HM Jr).
Raven Press, 51-79.
}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
