% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_ma2.R
\name{ma2}
\alias{ma2}
\title{Ichise Multilinear Analysis 2}
\usage{
ma2(
  t_tac,
  tac,
  input,
  weights = NULL,
  inpshift = 0,
  vB = 0,
  frameStartEnd = NULL
)
}
\arguments{
\item{t_tac}{Numeric vector of times for each frame in minutes. We use the time halfway through the frame as well as a
zero. If a time zero frame is not included, it will be added.}

\item{tac}{Numeric vector of radioactivity concentrations in the target tissue for each frame. We include zero at time
zero: if not included, it is added.}

\item{input}{Data frame containing the blood, plasma, and parent fraction concentrations over time.  This can be generated
using the \code{blood_interp} function.}

\item{weights}{Optional. Numeric vector of the weights assigned to each frame in the fitting. We include zero at time zero:
if not included, it is added. If not specified, uniform weights will be used.}

\item{inpshift}{Optional. The number of minutes by which to shift the timing of the input data frame forwards or backwards.
If not specified, this will be set to 0. This can be fitted using 1TCM or 2TCM.}

\item{vB}{Optional. The blood volume fraction.  If not specified, this will be ignored and assumed to be 0%. If specified, it
will be corrected for prior to parameter estimation using the following equation:
\deqn{C_{T}(t) = \frac{C_{Measured}(t) - vB\times C_{B}(t)}{1-vB}}}

\item{frameStartEnd}{Optional: This allows one to specify the beginning and final frame to use for modelling, e.g. c(1,20).
This is to assess time stability.}
}
\value{
A list with a data frame of the fitted parameters \code{out$par}, the model fit object \code{out$fit},
a dataframe containing the TACs of the data \code{out$tacs}, a dataframe containing the fitted values \code{out$fitvals},
the blood input data frame after time shifting \code{input}, a vector of the weights \code{out$weights},
the inpshift value used \code{inpshift} and the specified vB value \code{out$vB}.
}
\description{
Function to fit the MA2 of Ichise et al (2002) to data.
}
\examples{
data(pbr28)

t_tac <- pbr28$tacs[[2]]$Times / 60
tac <- pbr28$tacs[[2]]$FC
weights <- pbr28$tacs[[2]]$Weights

input <- blood_interp(
  pbr28$procblood[[2]]$Time / 60, pbr28$procblood[[2]]$Cbl_dispcorr,
  pbr28$procblood[[2]]$Time / 60, pbr28$procblood[[2]]$Cpl_metabcorr,
  t_parentfrac = 1, parentfrac = 1
)

fit1 <- ma2(t_tac, tac, input, weights)
fit2 <- ma2(t_tac, tac, input, weights, inpshift = 0.1, vB = 0.05)
}
\references{
Ichise M, Toyama H, Innis RB, Carson RE. Strategies to improve neuroreceptor parameter estimation by linear regression analysis. Journal of Cerebral Blood Flow & Metabolism. 2002 Oct 1;22(10):1271-81.
}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
