% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_miscfuncs.R
\name{maxpercres}
\alias{maxpercres}
\title{Maximum Percentage Residual}
\usage{
maxpercres(outputobject)
}
\arguments{
\item{outputobject}{The output of a kinetic model, including a fit object \code{outputobject$fit}}
}
\value{
The maximum percentage of the fitted value of a residual in the fit.
}
\description{
Function to determine the maximum percentage residual from the fitted value, i.e. the maximum
value of a residual as a percentage of its fitted value.  This is useful for obtaining estimates
of t*.
}
\examples{
refloganout <- refLogan(times, reftac, roitac, k2prime, tstarIncludedFrames = 9, weights=weights)
maxpercres(refloganout)

}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
