% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_pfmodels.R
\name{metab_exponential}
\alias{metab_exponential}
\title{Fit an Exponential Function for Modelling Parent Fraction.}
\usage{
metab_exponential(
  time,
  parentFraction,
  fit_ppf0 = FALSE,
  fit_delay = FALSE,
  lower = list(a = 0, b = 0, c = 0, ppf0 = 0.8, delay = -30),
  upper = list(a = 1, b = 1, c = 1, ppf0 = 1.1, delay = 30),
  multstart_lower = NULL,
  multstart_upper = NULL,
  multstart_iter = 100
)
}
\arguments{
\item{time}{Time in seconds.}

\item{parentFraction}{Measured values of parent fraction.}

\item{fit_ppf0}{Should the starting plasma parent fraction be fitted? Otherwise, it is set to 1. Defaults to FALSE.}

\item{fit_delay}{Should the delay of the plasma parent fraction be fitted? Otherwise, it is set to 0. Defaults to FALSE.}

\item{lower}{Named list of the lower limits.}

\item{upper}{Named list of the upper limits.}

\item{multstart_lower}{Named list of the lower starting limits.}

\item{multstart_upper}{Named list of the upper starting limits.}

\item{multstart_iter}{Number of fits to perform before deciding on an optimal.
Given as a single number, starting values are chosen by improved Latin
hypercube sampling rather than at random; given as one value per
parameter, they form a Cartesian grid.
See \code{\link[nls.multstart]{nls_multstart}}.}
}
\value{
An nls fit object.
}
\description{
This function fits the exponential function to parent fraction data.
}
\examples{
\dontrun{
pf <- bd_extract(blooddata, output = "parentFraction")
metab_exponential(pf$time, pf$parentFraction)
}
}
