% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_pfmodels.R
\name{metab_exponential_model}
\alias{metab_exponential_model}
\title{Exponential Function Model for Parent Fraction}
\usage{
metab_exponential_model(time, a, b, c, ppf0 = 1, delay = 0)
}
\arguments{
\item{time}{Time in seconds.}

\item{a}{Parameter A.}

\item{b}{Parameter B.}

\item{c}{Parameter C.}

\item{ppf0}{The starting point of the parent fraction curve.}

\item{delay}{The delay of the metabolism curve.}
}
\value{
Predicted values
}
\description{
This is the model function for fitting of the classic exponential function.
}
\examples{
metab_exponential_model(seq(0, 60*60, by=120), 0.02, 0, 0.001, 1, 0)
}
