% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_pfmodels.R
\name{metab_gamma_model}
\alias{metab_gamma_model}
\title{Gamma Function for Parent Fraction}
\usage{
metab_gamma_model(time, a, b, c, d, delay = 0)
}
\arguments{
\item{time}{Time in seconds.}

\item{a}{Parameter a. This parameter affects the ppf0 point.}

\item{b}{Parameter b.}

\item{c}{Shape parameter.}

\item{d}{Rate parameter.}

\item{delay}{The delay of the metabolism curve.}
}
\value{
Predicted values
}
\description{
This is the model function for fitting of the integrated gamma function for the parent fraction.
}
\examples{
metab_gamma_model(seq(0, 60 * 60, by = 120), 1.97, 708, 1, 0)
}
\references{
Naganawa M, Jacobsen LK, Zheng MQ, Lin SF, Banerjee A, Byon W, Weinzimmer D, Tomasi G, Nabulsi N, Grimwood S, Badura LL. Evaluation of the agonist PET radioligand [11C] GR103545 to image kappa opioid receptor in humans: Kinetic model selection, test–retest reproducibility and receptor occupancy by the antagonist PF-04455242. Neuroimage. 2014 Oct 1;99:69-79.
}
