% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_pfmodels.R
\name{metab_hill_model}
\alias{metab_hill_model}
\title{Hill Function Model for Parent Fraction}
\usage{
metab_hill_model(time, a, b, c, ppf0 = 0, delay = 1)
}
\arguments{
\item{time}{Time in seconds.}

\item{a}{Parameter A.}

\item{b}{Parameter B.}

\item{c}{Parameter C as the log10 of the 'classic' c parameter.}

\item{ppf0}{The starting point of the parent fraction curve.}

\item{delay}{The delay of the metabolism curve.}
}
\value{
Predicted values
}
\description{
This is the model function for fitting of the classic Hill function.
}
\examples{
metab_hill_model(seq(0, 60*60, by=120), 0.05, 2.6, 6.9, 1, 0)
}
