% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_pfmodels.R
\name{metab_invgamma_model}
\alias{metab_invgamma_model}
\title{Inverse Gamma Function Model for Parent Fraction}
\usage{
metab_invgamma_model(time, a, b, c, d, delay = 0)
}
\arguments{
\item{time}{Time in seconds.}

\item{a}{Parameter a. This parameter affects the ppf0 point.}

\item{b}{Parameter b.}

\item{c}{Shape parameter.}

\item{d}{Rate parameter.}

\item{delay}{The delay of the metabolism curve.}
}
\value{
Predicted values
}
\description{
This is the model function for fitting of the inverted integrated gamma function for the parent fraction.
}
\examples{
metab_invgamma_model(seq(0, 60 * 60, by = 120), 1, 0.95, 1.97, 708, 0)
}
