% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_mlloganplot.R
\name{mlLogan_tstar}
\alias{mlLogan_tstar}
\title{Tstar Finder: Multilinear Logan Plot}
\usage{
mlLogan_tstar(t_tac, lowroi, medroi, highroi, input, filename, inpshift = 0,
  frameStartEnd = NULL, gridbreaks = 2)
}
\arguments{
\item{t_tac}{Numeric vector of times for each frame in minutes. We use the time halfway through the frame as well as a
zero. If a time zero frame is not included, it will be added.}

\item{lowroi}{Numeric vector of radioactivity concentrations in a target tissue for each frame. This should be from a ROI with low binding.}

\item{medroi}{Numeric vector of radioactivity concentrations in a target tissue for each frame. This should be from a ROI with medium binding.}

\item{highroi}{Numeric vector of radioactivity concentrations in a target tissue for each frame. This should be from a ROI with high binding.}

\item{input}{Data frame containing the blood, plasma, and parent fraction concentrations over time.  This can be generated
using the \code{blood_interp} function.}

\item{filename}{The name of the output image: filename_Logan.jpeg}

\item{inpshift}{Optional. The number of minutes by which to shift the timing of the input data frame forwards or backwards.
If not specified, this will be set to 0. This can be fitted using 1TCM or 2TCM.}

\item{frameStartEnd}{Optional: This allows one to specify the beginning and final frame to use for modelling, e.g. c(1,20).
This is to assess time stability.}

\item{gridbreaks}{Optional. The size of the grid in the plots. Default: 2.}
}
\value{
Saves a jpeg of the plots as filename_mlLogan.jpeg
}
\description{
Function to identify where t* is for the multilinear Logan Plot.
}
\examples{
\dontrun{
mlLogan_tstar(t_tac, lowroi, medroi, highroi, input, filename='demonstration',
              inpshift = onetcmout$par$inpshift, vB = 0.05, gridbreaks=4)
}

}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
