% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_mlloganplot.R
\name{mlLoganplot}
\alias{mlLoganplot}
\title{Multilinear Logan Plot}
\usage{
mlLoganplot(
  t_tac,
  tac,
  input,
  tstar,
  weights = NULL,
  inpshift = 0,
  vB = 0,
  dur = NULL,
  frameStartEnd = NULL,
  timeStartEnd = NULL,
  tstar_type = "frames",
  tstarIncludedFrames = NULL
)
}
\arguments{
\item{t_tac}{Numeric vector of times for each frame in minutes. We use the time halfway through the frame as well as a
zero. If a time zero frame is not included, it will be added.}

\item{tac}{Numeric vector of radioactivity concentrations in the target tissue for each frame. We include zero at time
zero: if not included, it is added.}

\item{input}{Data frame containing the blood, plasma, and parent fraction concentrations over time.  This can be generated
using the \code{blood_interp} function.}

\item{tstar}{The t* specification for regression. If tstar_type="frames", 
this is the number of frames from the end to include (e.g., 10 means last 10 frames).
If tstar_type="time", this is the time point (in minutes) after which all frames 
with midpoints later than this time are included. This value can be estimated using \code{mlLogan_tstar}.}

\item{weights}{Optional. Numeric vector of the weights assigned to each frame in the fitting. We include zero at time zero:
if not included, it is added. If not specified, uniform weights will be used.}

\item{inpshift}{Optional. The number of minutes by which to shift the timing of the input data frame forwards or backwards.
If not specified, this will be set to 0. This can be fitted using 1TCM or 2TCM.}

\item{vB}{Optional. The blood volume fraction.  If not specified, this will be ignored and assumed to be 0%. If specified, it
will be corrected for prior to parameter estimation using the following equation:
\deqn{C_{T}(t) = \frac{C_{Measured}(t) - vB\times C_{B}(t)}{1-vB}}}

\item{dur}{Optional. Numeric vector of the time durations of the frames. If
not included, the integrals will be calculated using trapezoidal integration.}

\item{frameStartEnd}{Optional: This allows one to specify the beginning and final frame to use for modelling, e.g. c(1,20).
This can be used to assess time stability for example.}

\item{timeStartEnd}{Optional. This allows one to specify the beginning and end time point instead of defining the frame numbers using frameStartEnd. This function will restrict the model to all time frames whose t_tac is between the values, i.e. c(0,5) will select all frames with midtimes during the first 5 minutes.}

\item{tstar_type}{Either "frames" (default) or "time", specifying how to interpret tstar.}

\item{tstarIncludedFrames}{Deprecated. Use 'tstar' with 'tstar_type="frames"' instead.}
}
\value{
A list with a data frame of the fitted parameters \code{out$par}, the model fit object \code{out$fit},
a dataframe containing the TACs of the data \code{out$tacs}, a dataframe containing the fitted values \code{out$fitvals},
the blood input data frame after time shifting \code{input}, a vector of the weights \code{out$weights},
the inpshift value used \code{inpshift}, the specified vB value \code{out$vB}, and the specified tstarIncludedFrames
value \code{out$tstarIncludedFrames}.
}
\description{
Function to fit the multilinear Logan Plot of Turkheimer et al (2003) to data.
}
\examples{
data(pbr28)

t_tac <- pbr28$tacs[[2]]$Times / 60
tac <- pbr28$tacs[[2]]$FC
weights <- pbr28$tacs[[2]]$Weights
dur <- pbr28$tacs[[2]]$Duration/60

input <- blood_interp(
  pbr28$procblood[[2]]$Time / 60, pbr28$procblood[[2]]$Cbl_dispcorr,
  pbr28$procblood[[2]]$Time / 60, pbr28$procblood[[2]]$Cpl_metabcorr,
  t_parentfrac = 1, parentfrac = 1
)

fit1 <- mlLoganplot(t_tac, tac, input, 10, weights)
fit2 <- mlLoganplot(t_tac, tac, input, 10, weights, inpshift = 0.1, vB = 0.05)
fit3 <- mlLoganplot(t_tac, tac, input, 10, weights, inpshift = 0.1, vB = 0.05, dur=dur)
}
\references{
Turkheimer FE, Aston JA, Banati RB, Riddell C, Cunningham VJ. A linear wavelet filter for parametric imaging with dynamic PET. IEEE transactions on medical imaging. 2003 Mar;22(3):289-301.
}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
