% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_mrtm2.R
\name{mrtm2_tstar}
\alias{mrtm2_tstar}
\title{Tstar Finder: MRTM2}
\usage{
mrtm2_tstar(t_tac, reftac, lowroi, medroi, highroi, k2prime,
  filename = NULL, frameStartEnd = NULL, gridbreaks = 2)
}
\arguments{
\item{t_tac}{Numeric vector of times for each frame in minutes. We use the time halfway through the frame as well as a
zero. If a time zero frame is not included, it will be added.}

\item{reftac}{Numeric vector of radioactivity concentrations in the reference tissue for each frame.}

\item{lowroi}{Numeric vector of radioactivity concentrations in a target tissue for each frame. This should be from a ROI with low binding.}

\item{medroi}{Numeric vector of radioactivity concentrations in a target tissue for each frame. This should be from a ROI with medium binding.}

\item{highroi}{Numeric vector of radioactivity concentrations in a target tissue for each frame. This should be from a ROI with high binding.}

\item{k2prime}{Value of k2prime to be used for the fitting, i.e. the average tissue-to-plasma clearance rate. This can be
obtained from another model, such as MRTM1, SRTM or set at a specified value. If using SRTM to estimate this value, it is equal to k2 / R1.}

\item{filename}{The name of the output image: filename_mrtm1.jpeg}

\item{frameStartEnd}{Optional: This allows one to specify the beginning and final frame to use for modelling, e.g. c(1,20).
This is to assess time stability.}

\item{gridbreaks}{Optional. The size of the grid in the plots. Default: 2.}
}
\value{
Saves a jpeg of the plots as filename_mrtm2.jpeg
}
\description{
Function to identify where t* is for MRTM2. Remember that this is unnecessary if the kinetics in the target tissue can be described by a one
tissue compartment model (like SRTM).
}
\examples{
\dontrun{
mrtm2_tstar(t_tac, reftac, taclow, tacmed, tachigh, 'demonstration')
}

}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
