% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_1tcm.R
\name{onetcm_fitDelay_model}
\alias{onetcm_fitDelay_model}
\title{Model: One Tissue Compartment Model with Delay}
\usage{
onetcm_fitDelay_model(t_tac, input, K1, k2, inpshift, vB)
}
\arguments{
\item{t_tac}{Numeric vector of times for each frame in minutes. We use the time halfway through the frame as well as a zero.}

\item{input}{Data frame containing the blood, plasma, and parent fraction concentrations over time.  This can be generated
using the \code{blood_interp} function.}

\item{K1}{Parameter value for K1}

\item{k2}{Parameter value for k2}

\item{inpshift}{Parameter value for inpshift, the delay.}

\item{vB}{Parameter value for vB}
}
\value{
A numeric vector of the predicted values of the TAC in the target region.
}
\description{
This is the One Tissue Compartment Model model itself by which predicted values are generated, which includes fitting of the
delay, inpshift.
}
\examples{
\dontrun{
onetcm_fitDelay_model(t_tac, input, K1=0.1, k2=0.08, inpshift = 0.1, vB=0.05)
}

}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
