% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_methods.R
\name{plot.blooddata}
\alias{plot.blooddata}
\title{Plot blood data}
\usage{
\method{plot}{blooddata}(x, ...)
}
\arguments{
\item{x}{The \code{blooddata} object.}

\item{...}{Additional optional arguments.}
}
\description{
Function to plot the \code{blooddata} object.
}
\details{
This function provides an idea of how the blood data and their fitted data look.
}
\examples{
\dontrun{
plot(blooddata)
}

}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
