% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_methods.R
\name{plot.interpblood}
\alias{plot.interpblood}
\title{Plot interpolated blood}
\usage{
\method{plot}{interpblood}(x, ...)
}
\arguments{
\item{x}{The interpolated blood \code{input} object.}

\item{...}{Additional optional arguments.}
}
\description{
Function to plot the interpolated blood input (\code{interpblood} object).
}
\details{
This function provides an idea of how the input object looks.
}
\examples{
\dontrun{
input <- bd_create_input(blooddata)
plot(input)
}

}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
