% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_methods.R
\name{plot.kinfit}
\alias{plot.kinfit}
\title{Plot kinetic model fit}
\usage{
\method{plot}{kinfit}(x, ...)
}
\arguments{
\item{x}{The output object of the model fitting procedure.}

\item{...}{Additional optional arguments.}
}
\description{
Function to plot the output of a kinetic model.
}
\details{
This function uses the \code{out$model} name to call the correct function to plot the model fit.
}
\examples{
\dontrun{
loganout <- Loganplot(t_tac, tac, input, 10, weights)
plot(loganout)

srtmout <- srtm(t_tac, reftac, roitac)
plot(srtmout, roiname = 'Putamen', refname = 'Cerebellum')
}

}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
