% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_1tcm.R
\name{plot_1tcmfit}
\alias{plot_1tcmfit}
\title{Plot: One Tissue Compartment Model}
\usage{
plot_1tcmfit(onetcmout, roiname = NULL)
}
\arguments{
\item{onetcmout}{The output object of the 1TCM fitting procedure.}

\item{roiname}{Optional. The name of the Target Region to see it on the plot.}
}
\value{
A ggplot2 object of the plot.
}
\description{
Function to visualise the fit of the One Tissue Compartment Model to data.
}
\examples{

data(pbr28)

t_tac <- pbr28$tacs[[2]]$Times/60
tac <- pbr28$tacs[[2]]$FC
weights <- pbr28$tacs[[2]]$Weights

input <- blood_interp(
  pbr28$procblood[[2]]$Time/60 , pbr28$procblood[[2]]$Cbl_dispcorr,
  pbr28$procblood[[2]]$Time /60 , pbr28$procblood[[2]]$Cpl_metabcorr,
  t_parentfrac = 1, parentfrac = 1 )

fit <- onetcm(t_tac, tac, input, weights, inpshift=0.1, vB=0.05)

plot_1tcmfit(fit)

}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
