% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_2tcm.R
\name{plot_2tcmfit}
\alias{plot_2tcmfit}
\title{Plot: Two Tissue Compartment Model}
\usage{
plot_2tcmfit(twotcmout, roiname)
}
\arguments{
\item{twotcmout}{The output object of the 2TCM fitting procedure.}

\item{roiname}{Optional. The name of the Target Region to see it on the plot.}
}
\value{
A ggplot2 object of the plot.
}
\description{
Function to visualise the fit of the Two Tissue Compartment Model to data.
}
\examples{
plot_2tcmfit(twotcmout)

}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}

