% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_blooddata.R
\name{plot_blooddata}
\alias{plot_blooddata}
\title{Plot a blooddata object}
\usage{
plot_blooddata(blooddata, startTime = 0, stopTime = NULL,
  interpPoints = 6000, colours = c("#ff321b", "#1bff7e", "#501bff",
  "#ffa81b"))
}
\arguments{
\item{blooddata}{A blooddata object, to which all the desired fits have been applied and added.}

\item{startTime}{The starting time for the interpolation. Defaults to zero. If, after application of the TimeShift value in the blooddata object, the startTime is still after zero, it will be set to zero.}

\item{stopTime}{The end time for the interpolation. Defaults to the maximum measured time.}

\item{interpPoints}{The number of points to interpolate over between the start and stop times.}

\item{colours}{The colours for the plot.}
}
\value{
A ggplot2 object
}
\description{
Plot a blooddata object
}
\examples{
\dontrun{
plot_blooddata(blooddata)
}
}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
