% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_feng_1tc_ref.R
\name{plot_feng_1tc_tacfit}
\alias{plot_feng_1tc_tacfit}
\title{Plot Feng input model convolved with a 1TC IRF fit}
\usage{
plot_feng_1tc_tacfit(feng_1tc_tac_fit_out)
}
\arguments{
\item{feng_1tc_tac_fit_out}{The model output fit object.}
}
\value{
A ggplot2 object of the plot.
}
\description{
This function fits the results of the feng_1tc_tac model. For this
particular plotting function, I've added little crosses on the TAC
points in case the interpolation misses between points. The specific
values are important for SRTM and FRTM which use the "raw" reference
TAC, so extreme values at the exact times of the data points can
be problematic.
}
\examples{
data(simref)

t_tac <- simref$tacs[[2]]$Times
tac <- simref$tacs[[2]]$Reference
weights <- simref$tacs[[2]]$Weights

fit <- feng_1tc_tac(t_tac, tac, weights)

plot_feng_1tc_tacfit(fit)
}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
