% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_frtm.R
\name{plot_frtmfit}
\alias{plot_frtmfit}
\title{Plot: Full Reference Tissue Model}
\usage{
plot_frtmfit(frtmout, roiname = NULL, refname = NULL)
}
\arguments{
\item{frtmout}{The output object of the FRTM fitting procedure.}

\item{roiname}{Optional. The name of the Target Region to see it on the plot.}

\item{refname}{Optional. The name of the Reference Region to see it on the plot.}
}
\value{
A ggplot2 object of the plot.
}
\description{
Function to visualise the fit of the FRTM model to data.
}
\examples{
data(simref)

t_tac <- simref$tacs[[2]]$Times
reftac <- simref$tacs[[2]]$Reference
roitac <- simref$tacs[[2]]$ROI1
weights <- simref$tacs[[2]]$Weights

fit <- frtm(t_tac, reftac, roitac, weights, frameStartEnd = c(1, 33))

plot_frtmfit(fit)
}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
