% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_bloodfuncs.R
\name{plot_inptac_timings}
\alias{plot_inptac_timings}
\title{Plot the Timings of the TAC and Arterial Input Function}
\usage{
plot_inptac_timings(t_tac, tac, input, inpshift, zoomTime = 5)
}
\arguments{
\item{t_tac}{Numeric vector of times for each frame in minutes. We use the time halfway through the frame as well as a
zero. If a time zero frame is not included, it will be added.}

\item{tac}{Numeric vector of radioactivity concentrations in the target tissue for each frame. We include zero at time
zero: if not included, it is added.}

\item{input}{Data frame containing the blood, plasma, and parent fraction concentrations over time.  This can be generated
using the \code{blood_interp} function.}

\item{inpshift}{The number of minutes by which the times of the input data frame should be adjusted.}

\item{zoomTime}{The number of minutes to show for the close-up of the match between the TAC and AIF.}
}
\value{
A ggplot2 object of the plot.
}
\description{
Function to compare the timings of the the TAC and the Arterial Input Function (AIF) to see whether they are aligned.
Can be used for assessing the fit of the delay (\code{inpshift}).
}
\examples{
plot_inptac_timings(t_tac, tac, input, inpshift=0.25, zoomTime=5)

}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
