% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_loganplot.R
\name{plot_Loganfit}
\alias{plot_Loganfit}
\title{Plot: Logan Plot}
\usage{
plot_Loganfit(loganout, roiname = NULL)
}
\arguments{
\item{loganout}{The output object of the Logan Plot fitting procedure.}

\item{roiname}{Optional. The name of the Target Region to see it on the plot.}
}
\value{
A ggplot2 object of the plot.
}
\description{
Function to visualise the fit of the Logan Plot model to data.
}
\examples{
plot_Loganfit(loganoutout)

}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}

