% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_ma1.R
\name{plot_ma1fit}
\alias{plot_ma1fit}
\title{Plot: Ichise Multilinear Analysis 1}
\usage{
plot_ma1fit(ma1out, roiname = NULL)
}
\arguments{
\item{ma1out}{The output object of the MA1 fitting procedure.}

\item{roiname}{Optional. The name of the Target Region to see it on the plot.}
}
\value{
A ggplot2 object of the plot.
}
\description{
Function to visualise the fit of the MA1 model to data.
}
\examples{
plot_ma1fit(ma1out)

}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
