% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_ma2.R
\name{plot_ma2fit}
\alias{plot_ma2fit}
\title{Plot: Ichise Multilinear Analysis 2}
\usage{
plot_ma2fit(ma2out, roiname = NULL)
}
\arguments{
\item{ma2out}{The output object of the MA2 fitting procedure.}

\item{roiname}{Optional. The name of the Target Region to see it on the plot.}
}
\value{
A ggplot2 object of the plot.
}
\description{
Function to visualise the fit of the MA2 model to data.
}
\examples{
plot_ma2fit(ma2out)

}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
