% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_mlloganplot.R
\name{plot_mlLoganfit}
\alias{plot_mlLoganfit}
\title{Plot: Multilinear Logan Plot}
\usage{
plot_mlLoganfit(loganout, roiname = NULL)
}
\arguments{
\item{roiname}{Optional. The name of the Target Region to see it on the plot.}

\item{mlloganout}{The output object of the multilinear Logan Plot fitting procedure.}
}
\value{
A ggplot2 object of the plot.
}
\description{
Function to visualise the fit of the multilinear Logan Plot model to data.
}
\examples{
plot_mlLoganfit(mlloganoutout)

}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}

