% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_mrtm1.R
\name{plot_mrtm1fit}
\alias{plot_mrtm1fit}
\title{Plot: MRTM1}
\usage{
plot_mrtm1fit(mrtm1out, roiname = NULL, refname = NULL)
}
\arguments{
\item{mrtm1out}{The output object of the mrtm1 fitting procedure.}

\item{roiname}{Optional. The name of the Target Region to see it on the plot.}

\item{refname}{Optional. The name of the Reference Region to see it on the plot.}
}
\value{
A ggplot2 object of the plot.
}
\description{
Function to visualise the fit of the MRTM1 model to data.
}
\examples{
data(simref)

t_tac <- simref$tacs[[2]]$Times
reftac <- simref$tacs[[2]]$Reference
roitac <- simref$tacs[[2]]$ROI1
weights <- simref$tacs[[2]]$Weights

fit <- mrtm1(t_tac, reftac, roitac, weights = weights)

plot_mrtm1fit(fit)
}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
