% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_nested_delay.R
\name{plot_nested_2tcm_delayfit}
\alias{plot_nested_2tcm_delayfit}
\title{Plot nested 2TCM delay fit}
\usage{
plot_nested_2tcm_delayfit(nested_out, roiname = NULL)
}
\arguments{
\item{nested_out}{Output from \code{nested_2tcm_delay}.}

\item{roiname}{Optional. Not used (for API consistency).}
}
\value{
A ggplot2 object.
}
\description{
Plot nested 2TCM delay fit
}
