% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_patlakplot.R
\name{plot_Patlakfit}
\alias{plot_Patlakfit}
\title{Plot: Patlak Plot}
\usage{
plot_Patlakfit(patlakout, roiname = NULL)
}
\arguments{
\item{patlakout}{The output object of the Patlak Plot fitting procedure.}

\item{roiname}{Optional. The name of the Target Region to see it on the plot.}
}
\value{
A ggplot2 object of the plot.
}
\description{
Function to visualise the fit of the Patlak Plot model to data.
}
\examples{

data(pbr28)

t_tac <- pbr28$tacs[[2]]$Times / 60
tac <- pbr28$tacs[[2]]$FC
weights <- pbr28$tacs[[2]]$Weights

input <- blood_interp(
  pbr28$procblood[[2]]$Time / 60, pbr28$procblood[[2]]$Cbl_dispcorr,
  pbr28$procblood[[2]]$Time / 60, pbr28$procblood[[2]]$Cpl_metabcorr,
  t_parentfrac = 1, parentfrac = 1
)

fit <- Patlakplot(t_tac, tac, input, 10, weights, inpshift = 0.1)

plot_Patlakfit(fit)
}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
