% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_refmllogan.R
\name{plot_refmlLoganfit}
\alias{plot_refmlLoganfit}
\title{Plot: Non-Invasive Multilinear Logan Plot}
\usage{
plot_refmlLoganfit(refmlloganout, roiname = NULL)
}
\arguments{
\item{refmlloganout}{The output object of the refmlLogan fitting procedure.}

\item{roiname}{Optional. The name of the Target Region to see it on the plot.}
}
\value{
A ggplot2 object of the plot.
}
\description{
Function to visualise the fit of the refmlLogan model to data.
}
\examples{
data(simref)

t_tac <- simref$tacs[[2]]$Times
reftac <- simref$tacs[[2]]$Reference
roitac <- simref$tacs[[2]]$ROI1
weights <- simref$tacs[[2]]$Weights

fit <- refmlLogan(t_tac, reftac, roitac, k2prime = 0.1, tstarIncludedFrames = 10, weights = weights)

plot_refmlLoganfit(fit)
}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
