% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_refPatlak.R
\name{plot_refPatlakfit}
\alias{plot_refPatlakfit}
\title{Plot: Patlak Reference Tissue Model}
\usage{
plot_refPatlakfit(refpatlakout, roiname = NULL)
}
\arguments{
\item{refpatlakout}{The output object of the refPatlak fitting procedure.}

\item{roiname}{Optional. The name of the Target Region to see it on the plot.}
}
\value{
A ggplot2 object of the plot.
}
\description{
Function to visualise the fit of the refPatlak model to data.
}
\examples{
# Note: Reference region models, and irreversible binding models, should not
# be used for PBR28 - this is just to demonstrate function

data(pbr28)

t_tac <- pbr28$tacs[[2]]$Times/60
reftac <- pbr28$tacs[[2]]$CBL
roitac <- pbr28$tacs[[2]]$STR
weights <- pbr28$tacs[[2]]$Weights

fit <- refPatlak(t_tac, reftac, roitac, tstarIncludedFrames=10, weights=weights)

plot_refPatlakfit(fit)

}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
