% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_miscfuncs.R
\name{plot_residuals}
\alias{plot_residuals}
\title{Plot Residuals of a Model}
\usage{
plot_residuals(outputobject)
}
\arguments{
\item{outputobject}{The output of a kinetic model, including a fit object \code{outputobject$fit}}
}
\value{
A ggplot2 object of the residual plot
}
\description{
Function to visualise the residuals of a model fit.  This function only works for model outputs for
which a fit object is available called \code{outputobject$fit}.
}
\examples{
# Note: Reference region models, and irreversible binding models, should not
# be used for PBR28 - this is just to demonstrate function

data(pbr28)

t_tac <- pbr28$tacs[[2]]$Times/60
reftac <- pbr28$tacs[[2]]$CBL
roitac <- pbr28$tacs[[2]]$STR
weights <- pbr28$tacs[[2]]$Weights

srtmout <- srtm(t_tac, reftac, roitac)

plot_residuals(srtmout)

}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
