% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_SIME.R
\name{plot_SIMEfit}
\alias{plot_SIMEfit}
\title{Plot: SIME}
\usage{
plot_SIMEfit(SIMEout)
}
\arguments{
\item{SIMEout}{The output object of the SIME fitting procedure.}
}
\value{
A ggplot2 object of the plot.
}
\description{
The black line of this plot is the total cost function for the simultaneous estimation and is the true
outcome for the SIME model. I have also included, in coloured points, the costs for the individual ROIs such that 
one can visualise whether the overall estimated Vnd value is overly influenced by one ROI in particular, or whether
one ROI shows a completely different pattern from the rest of the ROIs.
}
\details{
Function to visualise the fit of the SIME Model to data.
}
\examples{
plot_SIMEfit(SIMEout)

}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}

