% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_spline_ref.R
\name{plot_spline_tacfit}
\alias{plot_spline_tacfit}
\title{Plot Spline TAC fit}
\usage{
plot_spline_tacfit(spline_tac_fit_out)
}
\arguments{
\item{spline_tac_fit_out}{The model output fit object.}
}
\value{
A ggplot2 object of the plot.
}
\description{
This function plots the results of the spline_tac model with interpolated
fitted curve. Small crosses are shown on the measured TAC points to highlight
the exact measured values, which is important since the interpolated curve
may miss between-point extrema.
}
\examples{
data(simref)

t_tac <- simref$tacs[[2]]$Times
tac <- simref$tacs[[2]]$Reference
weights <- simref$tacs[[2]]$Weights

fit <- spline_tac(t_tac, tac, weights)

plot_spline_tacfit(fit)
}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
