% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_srtm2.R
\name{plot_srtm2fit}
\alias{plot_srtm2fit}
\title{Plot: Simplified Reference Tissue Model 2}
\usage{
plot_srtm2fit(srtm2out, roiname = NULL, refname = NULL)
}
\arguments{
\item{srtm2out}{The output object of the SRTM2 fitting procedure.}

\item{roiname}{Optional. The name of the Target Region to see it on the plot.}

\item{refname}{Optional. The name of the Reference Region to see it on the plot.}
}
\value{
A ggplot2 object of the plot.
}
\description{
Function to visualise the fit of the SRTM2 model to data.
}
\examples{
data(simref)

t_tac <- simref$tacs[[2]]$Times
reftac <- simref$tacs[[2]]$Reference
roitac <- simref$tacs[[2]]$ROI1
weights <- simref$tacs[[2]]$Weights

fit <- srtm2(t_tac, reftac, roitac, weights=weights)

plot_srtm2fit(fit)
}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
