% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_srtm.R
\name{plot_srtmfit}
\alias{plot_srtmfit}
\title{Plot: Simplified Reference Tissue Model}
\usage{
plot_srtmfit(srtmout, roiname, refname)
}
\arguments{
\item{srtmout}{The output object of the SRTM fitting procedure.}

\item{roiname}{Optional. The name of the Target Region to see it on the plot.}

\item{refname}{Optional. The name of the Reference Region to see it on the plot.}
}
\value{
A ggplot2 object of the plot.
}
\description{
Function to visualise the fit of the SRTM model to data.
}
\examples{
plot_srtmfit(srtmout)

}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}

