% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_srtm_v.R
\name{plot_srtmvfit}
\alias{plot_srtmvfit}
\title{Plot: Simplified Reference Tissue Model with Blood Volumes}
\usage{
plot_srtmvfit(srtmvout, roiname, refname)
}
\arguments{
\item{srtmvout}{The output object of the SRTM fitting procedure.}

\item{roiname}{Optional. The name of the Target Region to see it on the plot.}

\item{refname}{Optional. The name of the Reference Region to see it on the plot.}
}
\value{
A ggplot2 object of the plot.
}
\description{
Function to visualise the fit of the SRTM_V model to data.
}
\examples{
plot_srtmvfit(srtmvout)

}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
