% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_methods.R
\name{predict.blmod}
\alias{predict.blmod}
\title{Predict Blood Model Values}
\usage{
\method{predict}{blmod}(object, ...)
}
\arguments{
\item{object}{The output object of the model fitting procedure.}

\item{...}{Additional optional arguments.}
}
\description{
Method to predict values from blood models
}
\details{
This function uses the \code{class(model)} name to call the correct function to plot the model fit.
}
\examples{
blooddata <- pbr28$blooddata[[1]]
blooddata <- bd_blood_dispcor(blooddata)
blood <- bd_extract(blooddata, output = "Blood")
blood_fit <- blmod_splines(blood$time,
                           blood$activity,
                           Method = blood$Method)

predict(blood_fit)
}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
