% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_bloodmodels.R
\name{predict_blood_exp}
\alias{predict_blood_exp}
\title{Predicted values for tri-exponential fits}
\usage{
predict_blood_exp(object, newdata = NULL)
}
\arguments{
\item{object}{Tri-exponential fit.}

\item{newdata}{A new data list, including times to be predicted for.}
}
\value{
Model predictions
}
\description{
Predicts values for new times for tri-exponential blood fits.
}
\examples{
blooddata <- pbr28$blooddata[[1]]
blood <- bd_extract(blooddata, output = "AIF")
blood_fit <- blmod_exp(blood$time,
                           blood$aif,
                           Method = blood$Method)

predict_blood_exp(blood_fit, newdata=list(time=1:10))
}
