% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_bloodmodels.R
\name{predict_blood_feng}
\alias{predict_blood_feng}
\title{Predicted values for Feng fits}
\usage{
predict_blood_feng(object, newdata = NULL)
}
\arguments{
\item{object}{Feng fit.}

\item{newdata}{A new data list, including times to be predicted for.}
}
\value{
Model predictions
}
\description{
Predicts values for new times for Feng blood fits.
}
\examples{
blooddata <- pbr28$blooddata[[1]]
blood <- bd_extract(blooddata, output = "AIF")
blood_fit <- blmod_feng(blood$time,
                           blood$aif,
                           Method = blood$Method)

predict_blood_feng(blood_fit, newdata=list(time=1:10))
}
