% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_bloodmodels.R
\name{predict_blood_fengconvplus}
\alias{predict_blood_fengconvplus}
\title{Predicted values for Fengconvplus fits}
\usage{
predict_blood_fengconvplus(object, newdata = NULL)
}
\arguments{
\item{object}{Fengconvplus fit.}

\item{newdata}{A new data list, including times to be predicted for.}
}
\value{
Model predictions
}
\description{
Predicts values for new times for Fengconvplus blood fits.
}
\examples{
\dontrun{
blooddata <- pbr28$blooddata[[1]]
blood <- bd_extract(blooddata, output = "AIF")
blood_fit <- blmod_fengconvplus(blood$time,
                           blood$aif,
                           Method = blood$Method)

predict_blood_fengconvplus(blood_fit, newdata=list(time=1:10))
}
}
