% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_bloodmodels.R
\name{predict_blood_splines}
\alias{predict_blood_splines}
\title{Predicted values for blood splines}
\usage{
predict_blood_splines(object, newdata = NULL)
}
\arguments{
\item{object}{Blood splines fit.}

\item{newdata}{A new data list, including times to be predicted for.}
}
\value{
Model predictions
}
\description{
Predicts values for new times for blood splines fits.
}
\examples{
blooddata <- pbr28$blooddata[[1]]
blood <- bd_extract(blooddata, output = "Blood")
blood_fit <- blmod_splines(blood$time,
                           blood$activity,
                           Method = blood$Method)

predict_blood_splines(blood_fit, newdata=list(time=1:10))
}
