% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_nested_utils.R
\name{print.nested_kinfit_plots}
\alias{print.nested_kinfit_plots}
\title{Print a paginated set of nested model plots}
\usage{
\method{print}{nested_kinfit_plots}(x, ...)
}
\arguments{
\item{x}{A \code{nested_kinfit_plots} object.}

\item{...}{Passed on to the print method of each plot.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a paginated set of nested model plots
}
