% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_reflogan.R
\name{refLogan_tstar}
\alias{refLogan_tstar}
\title{Tstar Finder: Non-Invasive Logan Plot}
\usage{
refLogan_tstar(t_tac, reftac, lowroi, medroi, highroi, k2prime, filename,
  gridbreaks = 2)
}
\arguments{
\item{t_tac}{Numeric vector of times for each frame in minutes. We use the time halfway through the frame as well as a
zero. If a time zero frame is not included, it will be added.}

\item{reftac}{Numeric vector of radioactivity concentrations in the reference tissue for each frame.}

\item{lowroi}{Numeric vector of radioactivity concentrations in a target tissue for each frame. This should be from a ROI with low binding.}

\item{medroi}{Numeric vector of radioactivity concentrations in a target tissue for each frame. This should be from a ROI with medium binding.}

\item{highroi}{Numeric vector of radioactivity concentrations in a target tissue for each frame. This should be from a ROI with high binding.}

\item{k2prime}{Value of k2prime to be used for the fitting, i.e. the average tissue-to-plasma clearance rate. This can be
obtained from another model, or set at a specified value. If using SRTM to estimate this value, it is equal to k2 / R1.}

\item{filename}{The name of the output image: filename_refLogan.jpeg}

\item{gridbreaks}{Optional. The size of the grid in the plots. Default: 2.}
}
\value{
Saves a jpeg of the plots as filename_refLogan.jpeg
}
\description{
Function to identify where t* is for the non-invasive Logan plot.
}
\examples{
refLogan_tstar(t_tac, reftac, taclow, tacmed, tachigh, k2prime = k2prime, 'demonstration')

}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
