% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_refPatlak.R
\name{refPatlak_tstar}
\alias{refPatlak_tstar}
\title{Tstar Finder: Patlak Reference Tissue Model}
\usage{
refPatlak_tstar(
  t_tac,
  reftac,
  lowroi,
  medroi,
  highroi,
  filename = NULL,
  frameStartEnd = NULL,
  timeStartEnd = NULL,
  gridbreaks = 2
)
}
\arguments{
\item{t_tac}{Numeric vector of times for each frame in minutes. We use the time halfway through the frame as well as a
zero. If a time zero frame is not included, it will be added.}

\item{reftac}{Numeric vector of radioactivity concentrations in the reference tissue for each frame.}

\item{lowroi}{Numeric vector of radioactivity concentrations in a target tissue for each frame. This should be from a ROI with low binding.}

\item{medroi}{Numeric vector of radioactivity concentrations in a target tissue for each frame. This should be from a ROI with medium binding.}

\item{highroi}{Numeric vector of radioactivity concentrations in a target tissue for each frame. This should be from a ROI with high binding.}

\item{filename}{The name of the output image: filename_refPatlak.jpeg}

\item{frameStartEnd}{Optional: This allows one to specify the beginning and final frame to use for modelling, e.g. c(1,20).
This can be used to assess time stability for example.}

\item{timeStartEnd}{Optional. This allows one to specify the beginning and end time point instead of defining the frame numbers using frameStartEnd. This function will restrict the model to all time frames whose t_tac is between the values, i.e. c(0,5) will select all frames with midtimes during the first 5 minutes.}

\item{gridbreaks}{Optional. The size of the grid in the plots. Default: 2.}
}
\value{
Saves a jpeg of the plots as filename_refPatlak.jpeg
}
\description{
Function to identify where t* is for the Patlak Reference Tissue Model.
}
\examples{
\dontrun{
refPatlak_tstar(t_tac, reftac, taclow, tacmed, tachigh, "demonstration")
}

}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
