% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_telemetry.R
\name{send_telemetry}
\alias{send_telemetry}
\title{Sending telemetry data}
\usage{
send_telemetry(telemetry_json_data, url = "http://54.144.240.214/kinfitr/")
}
\arguments{
\item{telemetry_json_data}{The JSON data content to be sent}

\item{url}{The url endpoint}
}
\value{
The status code
}
\description{
Sending telemetry data to the URL endpoint
}
