% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_SIME.R
\name{SIME}
\alias{SIME}
\title{Simultaneous Estimation of Non-Displaceable Binding (SIME)}
\usage{
SIME(
  t_tac,
  tac,
  region,
  input,
  Vndgrid,
  weights = NULL,
  roiweights = NULL,
  inpshift = 0,
  vB = NULL,
  twotcmstart = NULL,
  frameStartEnd = NULL,
  timeStartEnd = NULL,
  k2.start = 0.1,
  k2.lower = 0,
  k2.upper = 0.5,
  k3.start = 0.1,
  k3.lower = 0,
  k3.upper = 0.5,
  k4.start = 0.1,
  k4.lower = 0,
  k4.upper = 0.5,
  success_cutoff = 0.5
)
}
\arguments{
\item{t_tac}{Numeric vector of times for each frame in minutes (repeated for
each region). We use the time halfway through the frame as well as a zero.
If a time zero frame is not included, it will be added.}

\item{tac}{Numeric vector of radioactivity concentrations for each frame,
stacked across regions.}

\item{region}{Character vector identifying the region for each observation.}

\item{input}{Data frame containing the blood, plasma, and parent fraction
concentrations over time.  This can be generated using the
\code{blood_interp} function.}

\item{Vndgrid}{The grid of Vnd values which will be tested to see which one
has the best fit.}

\item{weights}{Optional. Numeric vector of the weights assigned to each frame
in the fitting. Can be the same length as \code{tac} (one weight per
observation) or the same length as the number of frames per region (weights
are recycled for each region). If not specified, uniform weights will be
used.}

\item{roiweights}{Optional. Named numeric vector of the weights assigned to
each ROI in the fitting, keyed by region name (e.g.,
\code{c(FC=1, STR=0.8)}). If unnamed, matched by position against unique
regions. If not specified, uniform weights will be used.}

\item{inpshift}{Optional. The number of minutes by which to shift the timing
of the input data frame forwards or backwards. If not specified, this will
be set to 0. This can be fitted using 1TCM or 2TCM.}

\item{vB}{Optional. The blood volume fraction.  If not specified, this will
be set to 0.05. This can be fitted using 1TCM or 2TCM.}

\item{twotcmstart}{Optional. Region name string specifying which region to
use for fitting a 2TCM model to obtain starting parameters for k2, k3 and
k4 (e.g., \code{"FC"}). Best to use the largest ROI.}

\item{frameStartEnd}{Optional: This allows one to specify the beginning and
final frame to use for modelling, e.g. c(1,20). Applied per region.}

\item{timeStartEnd}{Optional. This allows one to specify the beginning and
end time point instead of defining the frame numbers using frameStartEnd.
This function will restrict the model to all time frames whose t_tac is
between the values, i.e. c(0,5) will select all frames with midtimes
during the first 5 minutes.}

\item{k2.start}{Optional. Starting parameter for fitting of k2. Default is
0.1.}

\item{k2.lower}{Optional. Lower bound for the fitting of k2. Default is 0.}

\item{k2.upper}{Optional. Upper bound for the fitting of k2. Default is 0.5.}

\item{k3.start}{Optional. Starting parameter for fitting of k3. Default is
0.1.}

\item{k3.lower}{Optional. Lower bound for the fitting of k3. Default is 0.}

\item{k3.upper}{Optional. Upper bound for the fitting of k3. Default is 0.5.}

\item{k4.start}{Optional. Starting parameter for fitting of k4. Default is
0.1.}

\item{k4.lower}{Optional. Lower bound for the fitting of k4. Default is 0.}

\item{k4.upper}{Optional. Upper bound for the fitting of k4. Default is 0.5.}

\item{success_cutoff}{Optional. Should values of Vnd for which a certain
proportion of ROIs failed be included? Default is 0.5, i.e. 50% should have
successfully fitted.}
}
\value{
A list with a data frame of the fitted parameter \code{out$par}, a
  long-format data frame containing the times, regions and TACs
  \code{out$tacs} (with columns Time, Region, Radioactivity), the mean cost
  values after fitting (after ROI weighting) \code{out$fitvals}, the ROI cost
  values after fitting (before ROI weighting) \code{out$roifits}, the blood
  input data frame after time shifting \code{input}, a vector of the weights
  \code{out$weights}, a vector of the ROI weights \code{out$roiweights}, the
  inpshift value used \code{inpshift} and the vB value used \code{out$vB},
  and the success cutoff \code{success_cutoff}.
}
\description{
Function to fit the SIME Model of Ogden et al (2015) to data to estimate Vnd
for a set of TACs using long-format input data.
}
\examples{
\dontrun{
data(pbr28)

sime_long <- pbr28$tacs[[2]] \%>\%
  dplyr::select(Times, Weights, FC:CBL) \%>\%
  dplyr::mutate(Times = Times / 60) \%>\%
  tidyr::pivot_longer(cols = -c(Times, Weights),
                      names_to = "region", values_to = "tac")

input <- blood_interp(
  pbr28$procblood[[2]]$Time / 60, pbr28$procblood[[2]]$Cbl_dispcorr,
  pbr28$procblood[[2]]$Time / 60, pbr28$procblood[[2]]$Cpl_metabcorr,
  t_parentfrac = 1, parentfrac = 1
)

Vndgrid <- seq(from = 0, to = 3, by = 0.5)
SIMEout <- SIME(sime_long$Times, sime_long$tac, sime_long$region,
  input, Vndgrid,
  weights = sime_long$Weights,
  inpshift = 0.1, vB = 0.05
)
}
}
\references{
Ogden RT, Zanderigo F, Parsey RV. Estimation of in vivo
  nonspecific binding in positron emission tomography studies without
  requiring a reference region. NeuroImage. 2015 Mar 31;108:234-42.
}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
