% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_SIME.R
\name{SIME_model}
\alias{SIME_model}
\title{Model: SIME}
\usage{
SIME_model(t_tac, input, Vnd, k2, k3, k4, vB)
}
\arguments{
\item{t_tac}{Numeric vector of times for each frame in minutes. We use the time halfway through the frame as well as a
zero.}

\item{input}{Data frame containing the blood, plasma, and parent fraction concentrations over time.  This can be generated
using the \code{blood_interp} function. It should already be shifted if a shift is desired, using \code{shift_timings}.}

\item{Vnd}{The specified Vnd value for the given ROI.}

\item{k2}{Parameter value for k2}

\item{k3}{Parameter value for k3}

\item{k4}{Parameter value for k4}

\item{vB}{Parameter value for vB}
}
\value{
A numeric vector of the predicted values of the TAC in the target region with the given Vnd.
}
\description{
This is the SIME Model model itself by which predicted values are generated.
}
\examples{
\dontrun{
SIME_model(t_tac, input, Vnd=5, k2=0.1, k3=0.05, k4=0.04, vB=0.05)
}

}
\references{
Ogden RT, Zanderigo F, Parsey RV. Estimation of in vivo nonspecific binding in positron emission tomography studies without requiring a reference region. NeuroImage. 2015 Mar 31;108:234-42.
}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
