% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_SIME.R
\name{SIMEroi}
\alias{SIMEroi}
\title{Cost Function: SIME}
\usage{
SIMEroi(t_tac, tac, input, Vnd, vB, weights, k2.start = 0.1, k2.lower = 0,
  k2.upper = 0.5, k3.start = 0.1, k3.lower = 0, k3.upper = 0.5,
  k4.start = 0.1, k4.lower = 0, k4.upper = 0.5)
}
\arguments{
\item{t_tac}{Numeric vector of times for each frame in minutes. We use the time halfway through the frame as well as a 
zero. If a time zero frame is not included, it will be added.}

\item{tac}{Numeric vector of radioactivity concentrations in the target tissue for each frame including a zero at time 
zero.}

\item{input}{Data frame containing the blood, plasma, and parent fraction concentrations over time.  This can be generated
using the \code{blood_interp} function. It should already be shifted if a shift is desired, using \code{shift_timings}.}

\item{Vnd}{The specified Vnd value to calculate the cost for for the given ROI.}

\item{vB}{Optional. The blood volume fraction.  If not specified, this will be set to 0.05. This can be fitted using 1TCM or 2TCM.}

\item{weights}{Optional. Numeric vector of the weights assigned to each frame in the fitting. We include zero at time zero: 
if not included, it is added. If not specified, uniform weights will be used.}

\item{k2.start}{Optional. Starting parameter for fitting of k2. Default is 0.1.}

\item{k2.lower}{Optional. Lower bound for the fitting of k2. Default is 0.}

\item{k2.upper}{Optional. Upper bound for the fitting of k2. Default is 0.5.}

\item{k3.start}{Optional. Starting parameter for fitting of k3. Default is 0.1.}

\item{k3.lower}{Optional. Lower bound for the fitting of k3. Default is 0.}

\item{k3.upper}{Optional. Upper bound for the fitting of k3. Default is 0.5.}

\item{k4.start}{Optional. Starting parameter for fitting of k4. Default is 0.1.}

\item{k4.lower}{Optional. Lower bound for the fitting of k4. Default is 0.}

\item{k4.upper}{Optional. Upper bound for the fitting of k4. Default is 0.5.}
}
\value{
If the fit converged for the given Vnd value, the function will return the weighted sum of squares of the 
residuals. If the fit did not converge for the given Vnd value, the function will return NA.
}
\description{
Function to obtain the cost for a given ROI and Vnd for a TAC.
}
\examples{
SIMEroi(t_tac, tac, input, Vnd=5, vB=0.05, weights = weights)

}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
\references{
Ogden RT, Zanderigo F, Parsey RV. Estimation of in vivo nonspecific binding in positron emission tomography studies without requiring a reference region. NeuroImage. 2015 Mar 31;108:234-42.
}

