% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_srtm2.R
\name{srtm2_model}
\alias{srtm2_model}
\title{Model: Simplified Reference Tissue Model 2}
\usage{
srtm2_model(t_tac, reftac, R1, k2prime, bp)
}
\arguments{
\item{t_tac}{Numeric vector of times for each frame in minutes. We use the time halfway through the frame as well as a zero.}

\item{reftac}{Numeric vector of radioactivity concentrations in the reference tissue for each frame.}

\item{R1}{Parameter value for R1}

\item{k2prime}{Parameter value for k2prime}

\item{bp}{Parameter value for bp}
}
\value{
A numeric vector of the predicted values of the TAC in the target region.
}
\description{
This is the SRTM2 model itself by which predicted values are generated.
}
\examples{
\dontrun{
srtm2_model(t_tac, reftac, R1 = 0.9, k2prime = 0.1, bp = 0.1)
}

}
\references{
Wu Y, Carson RE. Noise reduction in the simplified reference tissue
  model for neuroreceptor functional imaging. J Cereb Blood Flow Metab.
  2002;22:1440-1452.
}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
