% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_srtm.R
\name{srtm_model}
\alias{srtm_model}
\title{Model: Simplified Reference Tissue Model}
\usage{
srtm_model(t_tac, reftac, R1, k2, bp)
}
\arguments{
\item{t_tac}{Numeric vector of times for each frame in minutes. We use the time halfway through the frame as well as a zero.}

\item{reftac}{Numeric vector of radioactivity concentrations in the reference tissue for each frame.}

\item{R1}{Parameter value for R1}

\item{k2}{Parameter value for k2}

\item{bp}{Parameter value for bp}
}
\value{
A numeric vector of the predicted values of the TAC in the target region.
}
\description{
This is the SRTM model itself by which predicted values are generated.
}
\examples{
srtm_model(t_tac, reftac, R1=0.9, k2=0.1, bp=1.5)

}
\references{
Lammertsma AA, Hume SP. Simplified reference tissue model for PET receptor studies. Neuroimage. 1996 Dec 31;4(3):153-8.
}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
