% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_srtm_v.R
\name{srtm_v_model}
\alias{srtm_v_model}
\title{Model: Simplified Reference Tissue Model with Blood Volumes}
\usage{
srtm_v_model(t_tac, reftac, bloodtac, R1, k2, bp, vBr, vBt)
}
\arguments{
\item{t_tac}{Numeric vector of times for each frame in minutes. We use the time halfway through the frame as well as a zero.}

\item{reftac}{Numeric vector of radioactivity concentrations in the reference tissue for each frame.}

\item{bloodtac}{Numeric vector of radioactivity concentrations in the blood for each frame.}

\item{R1}{Parameter value for R1}

\item{k2}{Parameter value for k2}

\item{bp}{Parameter value for bp}

\item{vBr}{Parameter value for vBr}

\item{vBt}{Parameter value for vBt}
}
\value{
A numeric vector of the predicted values of the TAC in the target region.
}
\description{
This is the SRTM model itself by which predicted values are generated.
}
\examples{
\dontrun{
srtm_v_model(t_tac, reftac, R1 = 0.9, k2 = 0.1, bp = 1.5, vBr = 0.05, vBt = 0.05)
}

}
\references{
Tomasi, G., Edison, P., ... & Turkheimer, F. E. (2008). Novel
  reference region model reveals increased microglial and reduced vascular
  binding of 11C-(R)-PK11195 in patients with Alzheimer's disease. Journal of
  Nuclear Medicine, 49(8), 1249-1256.
}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
