% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_miscfuncs.R
\name{tidyinput_art}
\alias{tidyinput_art}
\title{Tidy Up for Models with Arterial Input}
\usage{
tidyinput_art(t_tac, tac, weights, frameStartEnd)
}
\arguments{
\item{t_tac}{Numeric vector of times for each frame in minutes.}

\item{tac}{Numeric vector of radioactivity concentrations in the target tissue for each frame.}

\item{weights}{Optional. Numeric vector of the weights assigned to each frame in the fitting.}

\item{frameStartEnd}{Optional. This allows one to specify the beginning and final frame to use for modelling, e.g. c(1,20).}
}
\value{
A list containing the tidied up t_tac, tac and weights ready for modelling.
}
\description{
Function to tidy up the input argument vectors for arterial input models.
}
\details{
This function i) adds uniform weights if weights are not specified, ii) checks that the
lengths of t_tac, tac and weights are of the same length, iii) shortens the vectors
if a frameStartEnd is specified, iv) Checks whether there are negative values in t_tac due to 
adjusting for delay iv) adds a zero frame if there is not one, and v) checks that
times are in minutes and not in seconds.
}
\examples{
tidyinput_art(t_tac, tac, weights)

}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
