% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_2tcm_macro.R
\name{twotcm_macro_model}
\alias{twotcm_macro_model}
\title{Model: Two Tissue Compartment Model with a Macroparameter Parameterisation}
\usage{
twotcm_macro_model(t_tac, input, K1, Vnd, BPp, k4, vB)
}
\arguments{
\item{t_tac}{Numeric vector of times for each frame in minutes. We use the time halfway through the frame as well as a zero.}

\item{input}{Data frame containing the blood, plasma, and parent fraction concentrations over time.  This can be generated
using the \code{blood_interp} function.}

\item{K1}{Parameter value for K1}

\item{Vnd}{Parameter value for Vnd}

\item{BPp}{Parameter value for BPp}

\item{k4}{Parameter value for k4}

\item{vB}{Parameter value for vB}
}
\value{
A numeric vector of the predicted values of the TAC in the target region.
}
\description{
This is the Two Tissue Compartment Model model itself by which predicted values are generated.
}
\examples{
\dontrun{
twotcm_macro_model(t_tac, input, K1 = 0.1, Vnd = 1, BPp = 2, k4 = 0.02, vB = 0.05)
}

}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
