% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_miscfuncs.R
\name{unit_convert}
\alias{unit_convert}
\title{Convert between different units of radioactivity}
\usage{
unit_convert(values, from_units = c("GBq", "MBq", "kBq", "Bq", "mBq",
  "uBq", "nBq", "pBq", "GCi", "MCi", "kCi", "Ci", "mCi", "uCi", "nCi",
  "pCi"), to_units = c("GBq", "MBq", "kBq", "Bq", "mBq", "uBq", "nBq",
  "pBq", "GCi", "MCi", "kCi", "Ci", "mCi", "uCi", "nCi", "pCi"))
}
\arguments{
\item{values}{Radioactivity values which should be converted. This can be a single value or a vector of values.}

\item{from_units}{The units from which the radioactivity should be converted.}

\item{to_units}{The units to which the radioactivity should be converted.}
}
\value{
The converted radioactivity values in the new units
}
\description{
Convert between different units of radioactivity
}
\examples{
unit_convert(1, "nCi", "Bq")
unit_convert(1:5, "nCi", "Bq")
}
