% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_blooddata.R
\name{update_blooddata}
\alias{update_blooddata}
\title{Update an old blooddata object to the new structure}
\usage{
update_blooddata(blooddata)
}
\arguments{
\item{blooddata}{An old blooddata object in the old structure based on the
old PET BIDS specification}
}
\value{
A blooddata object in the new structure based on the newer PET BIDS
specification
}
\description{
This is a convenience function for updating old blooddata to a new structure.
This is not the ideal way of doing this: better would be to re-create a
blooddata object from the data saved according to the new BIDS specification.
This way is faster, but keeps only the vital information.
}
\examples{
\dontrun{
update_blooddata(blooddata)
}
}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
