% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_weights.R
\name{weights_sqrtdurcount}
\alias{weights_sqrtdurcount}
\title{Create weights using durations and counts}
\usage{
weights_sqrtdurcount(t_start, t_end, tac, radioisotope = c("C11", "O15",
  "F18"), minweight = 0.7, weight_checkn = 5)
}
\arguments{
\item{t_start}{The starting times of the frames in minutes.}

\item{t_end}{The end times of the frames in minutes.}

\item{tac}{The time activity curve.}

\item{radioisotope}{The radioisotope.}

\item{minweight}{The minimum weight. Weights will be calculated between this
value and 1.}

\item{weight_checkn}{The number of values of the weights to check to make
sure that things haven't gone terribly wrong. It will check that this
number of weights are all above half of the maximum weight.}
}
\value{
The vector of weights.
}
\description{
This is a non-standard weighting scheme, which is designed to approximate the
weighting scheme used at Karolinska Institutet, which is still awaiting
publication. It is advised to calculate weights once for each PET
measurement, using a large region (e.g. whole brain). It uses the following
equation: sqrt(durations) / sqrt(counts).
}
\examples{
data(pbr28)
s1 <- pbr28$tacs[[1]]

# Assuming the data were not decay-corrected (they are)
weights_sqrtdurcount(
  s1$StartTime/60,
  (s1$StartTime + s1$Duration)/60,
  radioisotope = "C11",
  tac = s1$WB)
}
