% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_blooddata.R
\name{bd_addfitpars}
\alias{bd_addfitpars}
\title{Add model fit parameters to a blooddata object}
\usage{
bd_addfitpars(
  blooddata,
  modelname,
  fitpars,
  modeltype = c("Blood", "BPR", "parentFraction", "AIF")
)
}
\arguments{
\item{blooddata}{A blooddata object created using one of the create_blooddata_* functions.}

\item{modelname}{The name of the model function which produces fitted outcomes for given parameters.}

\item{fitpars}{The fitted parameters as a named list, data.frame or tibble.}

\item{modeltype}{The function which the model predicts. One of the following:
Blood for models of how the blood data should be described, BPR for models
of the blood-to-plasma ratio, parentFraction for models of metabolism, and
AIF for models of the arterial input function.}
}
\value{
A blooddata object with the model inserted.
}
\description{
This function adds the fitted parameters of a known model to the blooddata object.
}
\examples{
\dontrun{
blooddata <- bd_addfitpars(blooddata, hillfit_model, hillfit$pars, "parentFraction")
}
}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
