% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_bloodmodels.R
\name{blmod_exp}
\alias{blmod_exp}
\title{Blood Model: Exponential}
\usage{
blmod_exp(
  time,
  activity,
  Method = NULL,
  weights = NULL,
  fit_t0 = TRUE,
  fit_exp3 = TRUE,
  fit_peaktime = FALSE,
  fit_peakval = FALSE,
  peaktime_val = NULL,
  peakval_set = TRUE,
  lower = NULL,
  upper = NULL,
  start = NULL,
  multstart_lower = NULL,
  multstart_upper = NULL,
  multstart_iter = 100,
  Method_weights = TRUE,
  taper_weights = TRUE,
  weightscheme = 2,
  check_startpars = FALSE,
  expdecay_props = c(1/60, 0.1)
)
}
\arguments{
\item{time}{The time of each measurement in seconds}

\item{activity}{The radioactivity of each measurement}

\item{Method}{Optional. The method of collection, i.e. "Discrete" or
"Continuous"}

\item{weights}{Optional. Weights of each measurement.}

\item{fit_t0}{Should time point zero be fitted? If not, it is set to 0.
Default is TRUE.}

\item{fit_exp3}{Should the third exponential be fitted, or should a
bi-exponential model be used? Default is TRUE for a tri-exponential model.}

\item{fit_peaktime}{Should the time of the peak be fitted? Default is FALSE.
This is potentially useful for data where sampling frequency was low around
the peak.}

\item{fit_peakval}{Should the value of the peak be fitted? Default is FALSE.
This is potentially useful for data where sampling frequency was low around
the peak.}

\item{peaktime_val}{Optional. If \code{fit_peaktime} is FALSE, then this
parameter will define the peaktime. If \code{fit_peaktime} is FALSE and
this parameter is left as NULL, then the empirical measured peaktime will
be used.}

\item{peakval_set}{Optional. If \code{fit_peakval} is FALSE, then this
parameter defines the peakval as either the empirical measured peakval
(TRUE), or the combination of A+B+c (FALSE).}

\item{lower}{Optional. The lower limits of the fit. If left as NULL, they
will be given reasonable defaults (mostly 50\% of the starting parameters).}

\item{upper}{Optional. The upper limits of the fit. If left as NULL, they
will be given reasonable defaults (mostly 150\% of the starting
parameters).}

\item{start}{Optional. The starting parameters for the fit. If left as NULL,
they will be selected using \code{blmod_exp_startpars}.}

\item{multstart_lower}{Optional. The lower limits of the starting parameters.}

\item{multstart_upper}{Optional. The upper limits of the starting parameters.}

\item{multstart_iter}{The number of fits to perform with different starting
parameters. If set to 1, then the starting parameters will be used for a
single fit. Otherwise a single number selects the starting values by
improved Latin hypercube sampling over the multstart bounds rather than at
random, while one value per parameter forms a Cartesian grid.}

\item{Method_weights}{If no weights provided, should the weights be divided
by discrete and continuous samples equally (i.e. with more continuous
samples, the continuous samples each get less weight). Default is TRUE.}

\item{taper_weights}{If no weights provided, should the weights be tapered to
gradually trade off between the continuous and discrete samples after the
peak?}

\item{weightscheme}{If no weights provided, which weighting scheme should be
used before accommodating Method_divide and taper_weights? Options are: 1
= uniform weighting; 2 = time/AIF as used by Columbia PET Centre; 3 =
absolute value of activity (weights rise with activity, emphasising the
peak); 4 = 1/abs(activity) (weights fall with activity, the
variance-stabilising direction for Poisson-like noise). Default is 2.}

\item{check_startpars}{Optional. Return only the starting parameters. Useful
for debugging fits which do not work.}

\item{expdecay_props}{What proportions of the decay should be used for
choosing starting parameters for the exponential decay. Defaults to 1/60
and 1/10, i.e. start to 1/60, 1/60 to 1/10 and 1/10 to end. If fitting only
two exponentials, the second value will be used.}
}
\value{
A model fit including all of the individual parameters, fit details,
  and model fit object of class blood_exp.
}
\description{
Fit an exponential model to AIF data with the ability to model the peak. In
other words, this model fits a single NLS model which describes the rise and
the fall of the AIF. This approach is more flexible, but also more sensitive
to fitting failures.
}
\details{
This model fits a bi- or tri-exponential model to AIF data. This model can be
specified in a fairly large number of ways, depending primarily on the
quality of the input data. Reasonably conservative defaults are provided.

The model can fit the time zero point (\code{fit_t0}), otherwise the rise
starts at time point zero. It can fit two or three exponentials to the curve
after the peak (\code{fit_exp0}). When it comes to fitting the peak itself,
the time of the peak can be fit or set (\code{fit_peaktime}). When the
peaktime is set, it is either set to the time point of the maximal measured
value, or it can be set to another value using \code{peaktime_val}. For
fitting the peak value, it can be fit as a unique parameter or set using
\code{fit_peakval}. If it is set, then it can either be set to the largest
measured value, or it can be set to the fitted combination of \code{A+B+C},
i.e. the initial part of the decay after the peak.
}
\examples{
blooddata <- pbr28$blooddata[[1]]
blooddata <- bd_blood_dispcor(blooddata)
aif <- bd_extract(blooddata, output = "AIF")
blood_fit <- blmod_exp(aif$time,
                           aif$aif,
                           Method = aif$Method, multstart_iter = 1)
}
