% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_bloodmodels.R
\name{blmod_fengconv}
\alias{blmod_fengconv}
\title{Blood Model: FengConv}
\usage{
blmod_fengconv(
  time,
  activity,
  inftime = NULL,
  Method = NULL,
  weights = NULL,
  fit_t0 = TRUE,
  lower = NULL,
  upper = NULL,
  start = NULL,
  multstart_lower = NULL,
  multstart_upper = NULL,
  multstart_iter = 500,
  taper_weights = TRUE,
  check_startpars = FALSE,
  expdecay_props = c(0, 0.5)
)
}
\arguments{
\item{time}{The time of each measurement in seconds}

\item{activity}{The radioactivity of each measurement}

\item{inftime}{The infusion time during which the tracer is administered. It
is recommended to provide this value if known. Leaving this argument empty
will result in the inftime being fitted. Alternatively, two values can be
specified, e.g. inftime=c(25, 35) to fit the infusion time with specified
upper and lower limits if the known infusion time is approximate.}

\item{Method}{Optional. The method of collection, i.e. "Discrete" or
"Continuous"}

\item{weights}{Optional. Weights of each measurement.}

\item{fit_t0}{Should time point zero be fitted? If not, it is set to 0.
Default is TRUE.}

\item{lower}{Optional. The lower limits of the fit. If left as NULL, they
will be given reasonable defaults (mostly 10\% of the starting parameters).}

\item{upper}{Optional. The upper limits of the fit. If left as NULL, they
will be given reasonable defaults (mostly 20x the starting parameters).}

\item{start}{Optional. The starting parameters for the fit. If left as NULL,
they will be selected using \code{blmod_feng_startpars}.}

\item{multstart_lower}{Optional. The lower limits of the starting parameters.}

\item{multstart_upper}{Optional. The upper limits of the starting parameters.}

\item{multstart_iter}{The number of fits to perform with different starting
parameters. If set to 1, then the starting parameters will be used for a
single fit.}

\item{taper_weights}{Should the weights be tapered to gradually trade off
between the continuous and discrete samples after the peak?}

\item{check_startpars}{Optional. Return only the starting parameters. Useful
for debugging fits which do not work.}

\item{expdecay_props}{What proportions of the decay should be used for
choosing starting parameters for the exponential decay. Defaults to 0 and
0.5, i.e. the latter two exponentials, B, C, beta, gamma are estimated
using halfway to the end of the decay, and the beginning to halfway through
the decay. The first parameters, A and alpha, are estimated from the
ascent.}
}
\value{
A model fit including all of the individual parameters, fit details,
  and model fit object of class blood_fengconv.
}
\description{
Fits a Feng model, convolved with an infusion injection to AIF data. This
model is significantly more effective than the original Feng model in the
vast majority of applications. The infusion duration can be a little bit
tricky to fit. I recommend trying to find out the approximate duration of the
injection and including this as a fixed parameter for most stable
performance.
}
\examples{
blooddata <- pbr28$blooddata[[1]]
blooddata <- bd_blood_dispcor(blooddata)
aif <- bd_extract(blooddata, output = "AIF")
blood_fit <- blmod_fengconv(aif$time,
                           aif$aif,
                           Method = aif$Method,
                           multstart_iter = 1,
                           inftime = 22)
}
\references{
Wong, Koon-Pong, Sung-Cheng Huang, and Michael J. Fulham.
  "Evaluation of an input function model that incorporates the injection
  schedule in FDG-PET studies." 2006 IEEE Nuclear Science Symposium
  Conference Record. Vol. 4. IEEE, 2006. Tonietto, M. (2012). Methods and
  models for the characterization of arterial input function in dynamic PET
  studies. Master thesis.
}
