% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_bloodmodels.R
\name{blmod_fengconvplus}
\alias{blmod_fengconvplus}
\title{Blood Model: FengConv plus a rise term}
\usage{
blmod_fengconvplus(
  time,
  activity,
  inftime = NULL,
  Method = NULL,
  weights = NULL,
  fit_t0 = TRUE,
  lower = NULL,
  upper = NULL,
  start = NULL,
  multstart_lower = NULL,
  multstart_upper = NULL,
  multstart_iter = 500,
  Method_weights = TRUE,
  taper_weights = TRUE,
  weightscheme = 2,
  check_startpars = FALSE,
  expdecay_props = c(0, 0.5)
)
}
\arguments{
\item{time}{The time of each measurement in seconds}

\item{activity}{The radioactivity of each measurement}

\item{inftime}{The infusion time during which the tracer is administered. It
is recommended to provide this value if known. Leaving this argument empty
will result in the inftime being fitted. Alternatively, two values can be
specified, e.g. inftime=c(25, 35) to fit the infusion time with specified
upper and lower limits if the known infusion time is approximate.}

\item{Method}{Optional. The method of collection, i.e. "Discrete" or
"Continuous"}

\item{weights}{Optional. Weights of each measurement.}

\item{fit_t0}{Should time point zero be fitted? If not, it is set to 0.
Default is TRUE.}

\item{lower}{Optional. The lower limits of the fit. If left as NULL, they
will be given reasonable defaults (mostly 50\% of the starting parameters).}

\item{upper}{Optional. The upper limits of the fit. If left as NULL, they
will be given reasonable defaults (mostly 150\% of the starting
parameters).}

\item{start}{Optional. The starting parameters for the fit. If left as NULL,
they will be selected using \code{blmod_exp_startpars}.}

\item{multstart_lower}{Optional. The lower limits of the starting parameters.}

\item{multstart_upper}{Optional. The upper limits of the starting parameters.}

\item{multstart_iter}{The number of fits to perform with different starting
parameters. If set to 1, then the starting parameters will be used for a
single fit.}

\item{Method_weights}{If no weights provided, should the weights be divided
by discrete and continuous samples equally (i.e. with more continuous
samples, the continuous samples each get less weight). Default is TRUE.}

\item{taper_weights}{If no weights provided, should the weights be tapered to
gradually trade off between the continuous and discrete samples after the
peak?}

\item{weightscheme}{If no weights provided, which weighting scheme should be
used before accommodating Method_divide and taper_weights? 1 represents a
uniform weighting before accommodating Method_divide and taper_weights. 2
represents time/AIF as used by Columbia PET Centre. Default is 2.}

\item{check_startpars}{Optional. Return only the starting parameters. Useful
for debugging fits which do not work.}

\item{expdecay_props}{What proportions of the decay should be used for
choosing starting parameters for the exponential decay. Defaults to 0 and
0.5, i.e. the latter two exponentials, B, C, beta, gamma are estimated
using halfway to the end of the decay, and the beginning to halfway through
the decay. The first parameters, A and alpha, are estimated from the
ascent.}
}
\value{
A model fit including all of the individual parameters, fit details,
  and model fit object of class blood_fengconvplus.
}
\description{
Fits a Feng model, convolved with an infusion injection, and then with an
additional rise to an asymptote term, to AIF data. This model is a slight
modification of the fengconv model which just allows the model to be a little
bit more capable of catching a rising tail: the model was designed for
bolus+infusion paradigms - however it should be stated that FengConv is
usually pretty capable in these cases too. By bolus+infusion, I mean a bolus,
injected as a short infusion, followed by a long infusion. The infusion
duration can be a little bit tricky to fit. I recommend trying to find out
the approximate duration of the injection and including this as a fixed
parameter for most stable performance.
}
\examples{
\dontrun{
blooddata <- pbr28$blooddata[[1]]
blooddata <- bd_blood_dispcor(blooddata)
aif <- bd_extract(blooddata, output = "AIF")
blood_fit <- blmod_fengconvplus(aif$time,
                           aif$aif,
                           Method = aif$Method,
                           multstart_iter = 100)
}
}
