% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_bloodmodels.R
\name{blmod_splines}
\alias{blmod_splines}
\title{Blood Model: Splines}
\usage{
blmod_splines(
  time,
  activity,
  Method = NULL,
  weights = NULL,
  Method_weights = TRUE,
  taper_weights = TRUE,
  weightscheme = 2,
  bs_before = "cr",
  bs_after_c = "cr",
  bs_after_d = "cr",
  k_before = -1,
  k_after_c = -1,
  k_after_d = -1
)
}
\arguments{
\item{time}{The time of each measurement in seconds}

\item{activity}{The radioactivity of each measurement}

\item{Method}{Optional. The method of collection, i.e. "Discrete" or
"Continuous"}

\item{weights}{Optional. Weights of each measurement.}

\item{Method_weights}{If no weights provided, should the weights be divided
by discrete and continuous samples equally (i.e. with more continuous
samples, the continuous samples each get less weight). Default is TRUE.}

\item{taper_weights}{If no weights provided, should the weights be tapered to
gradually trade off between the continuous and discrete samples after the
peak?}

\item{weightscheme}{If no weights provided, which weighting scheme should be
used before accommodating Method_divide and taper_weights? 1 represents a
uniform weighting before accommodating Method_divide and taper_weights. 2
represents time/AIF as used by Columbia PET Centre. Default is 2.}

\item{bs_before}{Optional. Defines the basis function for the points before
the peak.}

\item{bs_after_c}{Optional. Defines the basis function for the continuous
points after the peak.}

\item{bs_after_d}{Optional. Defines the basis function for the discrete
points after the peak.}

\item{k_before}{Optional. Defines the dimension of the basis for the points
before the peak.}

\item{k_after_c}{Optional. Defines the dimension of the basis for the
continuous points after the peak.}

\item{k_after_d}{Optional. Defines the dimension of the basis for the
discrete points after the peak.}
}
\value{
A model fit including all of the individual models of class
  blood_splines.
}
\description{
Fits two or three (if both discrete and continuous) splines to blood or AIF
data to smooth it
}
\examples{
blooddata <- pbr28$blooddata[[1]]
blooddata <- bd_blood_dispcor(blooddata)
blood <- bd_extract(blooddata, output = "Blood")
blood_fit <- blmod_splines(blood$time,
                           blood$activity,
                           Method = blood$Method)
}
