% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_bloodmodels.R
\name{blmod_splines}
\alias{blmod_splines}
\title{Blood Model: Splines}
\usage{
blmod_splines(
  time,
  activity,
  Method = NULL,
  weights = NULL,
  bs_before = "cr",
  bs_after_c = "ad",
  bs_after_d = "cr",
  k_before = -1,
  k_after_c = -1,
  k_after_d = -1
)
}
\arguments{
\item{time}{The time of each measurement in seconds}

\item{activity}{The radioactivity of each measurement}

\item{Method}{Optional. The method of collection, i.e. "Discrete" or "Continuous"}

\item{weights}{Optional. Weights of each measurement.}

\item{bs_before}{Optional. Defines the basis function for the points before the peak.}

\item{bs_after_c}{Optional. Defines the basis function for the continuous points after the peak.}

\item{bs_after_d}{Optional. Defines the basis function for the discrete points after the peak.}

\item{k_before}{Optional. Defines the dimension of the basis for the points before the peak.}

\item{k_after_c}{Optional. Defines the dimension of the basis for the continuous points after the peak.}

\item{k_after_d}{Optional. Defines the dimension of the basis for the discrete points after the peak.}
}
\value{
A model fit including all of the individual models of class blood_splines.
}
\description{
Fits two or three (if both discrete and continuous) splines to blood or AIF data to smooth it
}
\examples{
blooddata <- create_blooddata_bids(pbr28$jsondata[[1]])
blooddata <- bd_blood_dispcor(blooddata)
blood <- bd_getdata(blooddata, output = "Blood")
blood_fit <- blmod_splines(blood$time,
                           blood$activity,
                           Method = blood$Method)
}
