% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_bloodmodels.R
\name{blmod_splines}
\alias{blmod_splines}
\title{Blood Model: Splines}
\usage{
blmod_splines(time, activity, Method = NULL, weights = NULL)
}
\arguments{
\item{time}{The time of each measurement in seconds}

\item{activity}{The radioactivity of each measurement}

\item{Method}{Optional. The method of collection, i.e. "Discrete" or "Continuous"}

\item{weights}{Optional. Weights of each measurement.}
}
\value{
A model fit including all of the individual models.
}
\description{
Fits two or three (if both discrete and continuous) splines to blood or AIF data to smooth it
}
\examples{
blooddata <- create_blooddata_bids(pbr28$jsondata[[1]])
blood <- bd_getdata(blooddata, output = "Blood")
blood_fit <- blmod_splines(blood$time,
                           blood$activity,
                           Method = blood$Method)
}
