% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_bloodmodels.R
\name{blmod_tidyinput}
\alias{blmod_tidyinput}
\title{Blood Model: Tidy inputs}
\usage{
blmod_tidyinput(time, activity, Method = NULL, weights = NULL)
}
\arguments{
\item{time}{The time of each measurement in seconds}

\item{activity}{The radioactivity of each measurement}

\item{Method}{Optional. The method of collection, i.e. "Discrete" or "Continuous"}

\item{weights}{Optional. Weights of each measurement.}
}
\value{
A blood data frame ready for modelling
}
\description{
Tidies the inputs to blood models ready for modelling
}
\examples{
blooddata <- create_blooddata_bids(pbr28$jsondata[[1]])
blood <- bd_getdata(blooddata, output = "Blood")
blood <- blmod_tidyinput(blood$time,
                           blood$activity,
                           Method = blood$Method)
}
