% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_bloodmodels.R
\name{blmod_triexp_model}
\alias{blmod_triexp_model}
\title{Tri-exponential model for fitting of arterial input functions}
\usage{
blmod_triexp_model(time, t0, peaktime, peakval, A, alpha, B, beta, C,
  gamma)
}
\arguments{
\item{time}{Time of each sample.}

\item{t0}{The delay time. This is the point at which the linear rise begins.}

\item{peaktime}{The time of the peak.}

\item{peakval}{The radioactivity value of the peak.}

\item{A}{The multiplier of the first exponential.}

\item{alpha}{The rate of the first exponential.}

\item{B}{The multiplier of the second exponential.}

\item{beta}{The rate of the second exponential.}

\item{C}{The multiplier of the third exponential.}

\item{gamma}{The rate of the third exponential.}
}
\value{
Model predictions
}
\description{
This is the model itself for the tri-exponential model of AIF decay, in which
both the rise and fall are modelled together in the same nonlinear model.
}
\examples{
blmod_triexp_model(1:100, 2, 5, 100, 80, 0.5, 15, 0.1, 5, 0.01)
}
