% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_bloodfuncs.R
\name{blood_weights_create}
\alias{blood_weights_create}
\title{Create Weights for Modelling Whole Blood and AIF Data}
\usage{
blood_weights_create(
  time,
  activity,
  Method = NULL,
  Method_weights = TRUE,
  taper_weights = TRUE,
  weightscheme = 2
)
}
\arguments{
\item{time}{The time of each measurement.}

\item{activity}{The radioactivity of each measurement.}

\item{Method}{Optional. The method of collection, i.e. "Discrete" or
"Continuous".}

\item{Method_weights}{Should the weights be divided by discrete and
continuous samples equally (i.e. with more continuous samples, the
continuous samples each get less weight). Default is TRUE.}

\item{taper_weights}{Should the weights be tapered to gradually trade off
between the continuous and discrete samples after the peak?}

\item{weightscheme}{Which weighting scheme should be used before
accommodating Method_divide and taper_weights? 1 represents a uniform
weighting before accommodating Method_divide and taper_weights. 2
represents time/AIF as used by Columbia PET Centre. Default is 2.}
}
\value{
A vector of model weights
}
\description{
Create Weights for Modelling Whole Blood and AIF Data
}
\examples{
blooddata <- pbr28$blooddata[[1]]
aif <- bd_extract(blooddata, output = "AIF")
blood_weights_create(aif$time, aif$aif, aif$Method)
}
