% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_blooddata.R
\name{create_blooddata_components}
\alias{create_blooddata_components}
\title{Create a blooddata object from data vectors}
\usage{
create_blooddata_components(
  Blood.Discrete.Values.time = NULL,
  Blood.Discrete.Values.activity = NULL,
  Plasma.Values.time = NULL,
  Plasma.Values.activity = NULL,
  Metabolite.Values.time = NULL,
  Metabolite.Values.parentFraction = NULL,
  Blood.Continuous.Values.time = NULL,
  Blood.Continuous.Values.activity = NULL,
  Blood.Continuous.DispersionConstant = NULL,
  Blood.Continuous.DispersionCorrected = TRUE,
  TimeShift = 0
)
}
\arguments{
\item{Blood.Discrete.Values.time}{Sample times in seconds.}

\item{Blood.Discrete.Values.activity}{Measured radioactivity in kBq/ml.}

\item{Plasma.Values.time}{In seconds. Sample times in seconds.}

\item{Plasma.Values.activity}{Measured radioactivity in kBq/ml.}

\item{Metabolite.Values.time}{Sample times in seconds.}

\item{Metabolite.Values.parentFraction}{Measured fraction (unitless).}

\item{Blood.Continuous.Values.time}{In seconds.}

\item{Blood.Continuous.Values.activity}{in kBq/ml.}

\item{Blood.Continuous.DispersionConstant}{External dispersion time constant
resulting from tubing in seconds.}

\item{Blood.Continuous.DispersionCorrected}{Boolean flag specifying whether
the continuous blood data have been dispersion-corrected.}

\item{TimeShift}{The extent to which all the times in the data should be
shifted (in seconds). Defaults to 0.}
}
\value{
a blooddata object
}
\description{
This function creates a blooddata object from data vectors. This function
creates data objects which contain the minimum amount of data for modelling.
Leave blank fields which do not exist.
}
\details{
Ideally, I recommend storing your data according to the BIDS specification
and using the bids_create_blooddata() command instead. Please refer to the PET
BIDS standard for further details about the inputs.
}
\examples{

blooddata <- pbr28$blooddata[[1]]

blooddata2 <- create_blooddata_components(
   Blood.Discrete.Values.time =
     blooddata$Data$Blood$Discrete$Values$time,
   Blood.Discrete.Values.activity =
     blooddata$Data$Blood$Discrete$Values$activity,
   Plasma.Values.time =
     blooddata$Data$Plasma$Values$time,
   Plasma.Values.activity =
     blooddata$Data$Plasma$Values$activity,
   Metabolite.Values.time =
     blooddata$Data$Metabolite$Values$time,
   Metabolite.Values.parentFraction =
     blooddata$Data$Metabolite$Values$parentFraction,
   Blood.Continuous.Values.time =
     blooddata$Data$Blood$Continuous$Values$time,
   Blood.Continuous.Values.activity =
     blooddata$Data$Blood$Continuous$Values$activity,
   Blood.Continuous.DispersionConstant =
     blooddata$Data$Blood$Continuous$DispersionConstant,
   Blood.Continuous.DispersionCorrected = FALSE,
   TimeShift = 0)

}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
