% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_miscfuncs.R
\name{fix_multstartpars}
\alias{fix_multstartpars}
\title{Check and fix multstart parameters}
\usage{
fix_multstartpars(start, lower, upper, multstart_iter, multstart_lower,
  multstart_upper)
}
\arguments{
\item{start}{Original starting values for without multstart}

\item{lower}{Lower fitting bounds}

\item{upper}{Upper fitting bounds}

\item{multstart_iter}{Iterations}

\item{multstart_lower}{Lower starting bounds}

\item{multstart_upper}{Upper starting bounds}
}
\value{
A list containing the parameters
}
\description{
Checks and fixes for inputted multstart parameters. Checks missing multstart boundaries,
checks if some parameters aren't to be tried with different starting parameters, and checks
the lengths of the multstart bounds
}
\examples{
start <- list(a = 1, b = 2, c = 3)
upper <- lapply(start, function(x) x+1)
lower <- lapply(start, function(x) x-1)
multstart_iter <- 5

multstart_lower <- NULL
multstart_upper <- NULL

fix_multstartpars(start, lower, upper, multstart_iter, multstart_lower, multstart_upper)

}
