% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_LEGA.R
\name{LEGA_tstar}
\alias{LEGA_tstar}
\title{Tstar Finder: Likelihood Estimation in Graphical Analysis (LEGA)}
\usage{
LEGA_tstar(
  t_tac,
  lowroi,
  medroi,
  highroi,
  input,
  filename = NULL,
  inpshift = 0,
  vB = 0,
  frameStartEnd = NULL,
  timeStartEnd = NULL,
  gridbreaks = 2
)
}
\arguments{
\item{t_tac}{Numeric vector of times for each frame in minutes. We use the time halfway through the frame as well as a
zero. If a time zero frame is not included, it will be added.}

\item{lowroi}{Numeric vector of radioactivity concentrations in a target tissue for each frame. This should be from a ROI with low binding.}

\item{medroi}{Numeric vector of radioactivity concentrations in a target tissue for each frame. This should be from a ROI with medium binding.}

\item{highroi}{Numeric vector of radioactivity concentrations in a target tissue for each frame. This should be from a ROI with high binding.}

\item{input}{Data frame containing the blood, plasma, and parent fraction concentrations over time.  This can be generated
using the \code{blood_interp} function.}

\item{filename}{The name of the output image: filename_LEGA.jpeg}

\item{inpshift}{Optional. The number of minutes by which to shift the timing of the input data frame forwards or backwards.
If not specified, this will be set to 0. This can be fitted using 1TCM or 2TCM.}

\item{vB}{Optional. The blood volume fraction.  If not specified, this will
be ignored and assumed to be 0%. If specified, it will be corrected for
prior to parameter estimation using the following equation: \deqn{C_{T}(t)
= \frac{C_{Measured}(t) - vB\times C_{B}(t)}{1-vB}}}

\item{frameStartEnd}{Optional: This allows one to specify the beginning and final frame to use for modelling, e.g. c(1,20).
This can be used to assess time stability for example.}

\item{timeStartEnd}{Optional. This allows one to specify the beginning and end time point instead of defining the frame numbers using frameStartEnd. This function will restrict the model to all time frames whose t_tac is between the values, i.e. c(0,5) will select all frames with midtimes during the first 5 minutes.}

\item{gridbreaks}{Optional. The size of the grid in the plots. Default: 2.}
}
\value{
Saves a jpeg of the plots as filename_LEGA.jpeg
}
\description{
Function to identify where t* is for LEGA.
}
\examples{
\dontrun{
LEGA_tstar(t_tac, lowroi, medroi, highroi, input,
  filename = "demonstration",
  inpshift = onetcmout$par$inpshift, frameStartEnd, gridbreaks = 4
)
}

}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
