% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_miscfuncs.R
\name{maxpercres}
\alias{maxpercres}
\title{Maximum Percentage Residual}
\usage{
maxpercres(outputobject)
}
\arguments{
\item{outputobject}{The output of a kinetic model, including a fit object \code{outputobject$fit}}
}
\value{
The maximum percentage of the fitted value of a residual in the fit.
}
\description{
Function to determine the maximum percentage residual from the fitted value, i.e. the maximum
value of a residual as a percentage of its fitted value.  This is useful for obtaining estimates
of t*.
}
\examples{

# Note: Reference region models, and irreversible binding models, should not
# be used for PBR28 - this is just to demonstrate function

data(pbr28)

t_tac <- pbr28$tacs[[2]]$Times/60
reftac <- pbr28$tacs[[2]]$CBL
roitac <- pbr28$tacs[[2]]$STR
weights <- pbr28$tacs[[2]]$Weights

refloganout <- refLogan(t_tac, reftac, roitac, 0.1, tstarIncludedFrames = 9)

maxpercres(refloganout)

}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
