% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_pfmodels.R
\name{metab_invgamma}
\alias{metab_invgamma}
\title{Fit the Inverted Gamma Function for Modelling Parent Fraction.}
\usage{
metab_invgamma(time, parentFraction, fit_ppf0 = FALSE,
  fit_delay = FALSE, lower = list(shape = 0, rate = 0, ppf0 = 0.8,
  delay = -30), upper = list(shape = 1000, rate = 1000, ppf0 = 1.1, delay
  = 30), multstart_lower = NULL, multstart_upper = NULL,
  multstart_iter = 100)
}
\arguments{
\item{time}{Time in seconds.}

\item{parentFraction}{Measured values of parent fraction.}

\item{fit_ppf0}{Should the starting plasma parent fraction be fitted? Otherwise, it is set to 1. Defaults to FALSE.}

\item{fit_delay}{Should the delay of the plasma parent fraction be fitted? Otherwise, it is set to 0. Defaults to FALSE.}

\item{lower}{Named list of the lower limits.}

\item{upper}{Named list of the upper limits.}

\item{multstart_lower}{Named list of the lower starting limits.}

\item{multstart_upper}{Named list of the upper starting limits.}

\item{multstart_iter}{Number of fits to perform before deciding on an optimal.}
}
\value{
An nls fit object.
}
\description{
This function fits the inverted gamma function to parent fraction data.
}
\examples{
\dontrun{
pf <- blooddata_getdata(blooddata, output="parentFraction")
metab_invgamma(pf$time, pf$parentFraction)
}
}
