% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_pfmodels.R
\name{metab_invgamma_model}
\alias{metab_invgamma_model}
\title{Inverse Gamma Function Model for Parent Fraction}
\usage{
metab_invgamma_model(time, shape, rate, ppf0 = 1, delay = 0)
}
\arguments{
\item{time}{Time in seconds.}

\item{shape}{Shape parameter.}

\item{rate}{Rate parameter.}

\item{ppf0}{The starting point of the parent fraction curve.}

\item{delay}{The delay of the metabolism curve.}
}
\value{
Predicted values
}
\description{
This is the model function for fitting of the inverse Gamma function for the parent fraction.
}
\examples{
metab_invgamma_model(seq(0, 60*60, by=120), 1.97, 708, 1, 0)
}
