% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_pfmodels.R
\name{metab_power}
\alias{metab_power}
\title{Fit a Power Function for Modelling Parent Fraction.}
\usage{
metab_power(
  time,
  parentFraction,
  fit_ppf0 = FALSE,
  fit_delay = FALSE,
  lower = list(a = 0, b = 1, c = 0, ppf0 = 0.8, delay = -30),
  upper = list(a = 1, b = 10, c = 5, ppf0 = 1.1, delay = 30),
  multstart_lower = NULL,
  multstart_upper = NULL,
  multstart_iter = 100
)
}
\arguments{
\item{time}{Time in seconds.}

\item{parentFraction}{Measured values of parent fraction.}

\item{fit_ppf0}{Should the starting plasma parent fraction be fitted? Otherwise, it is set to 1. Defaults to FALSE.}

\item{fit_delay}{Should the delay of the plasma parent fraction be fitted? Otherwise, it is set to 0. Defaults to FALSE.}

\item{lower}{Named list of the lower limits.}

\item{upper}{Named list of the upper limits.}

\item{multstart_lower}{Named list of the lower starting limits.}

\item{multstart_upper}{Named list of the upper starting limits.}

\item{multstart_iter}{Number of fits to perform before deciding on an optimal.
Starting values are chosen by improved Latin hypercube sampling rather
than at random. See \code{\link[nls.multstart]{nls_multstart}}.}
}
\value{
An nls fit object.
}
\description{
This function fits the power function to parent fraction data.
}
\examples{
\dontrun{
pf <- bd_getdata(blooddata, output = "parentFraction")
metab_power(pf$time, pf$parentFraction)
}
}
