% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledata.R
\docType{data}
\name{pbr28}
\alias{pbr28}
\title{PBR28 Test-Retest Data}
\format{A nested data frame with 24 rows and 10 variables:
\describe{
  \item{PET}{Descriptor of which PET measurement}
  \item{Subjname}{Descriptor of which individual was measured}
  \item{PETNo}{Descriptor of which PET number}
  \item{tacs}{Nested tibble of the TAC data. Consists of Times, in seconds; Weights, between 0 and 1; StartTime, in seconds; Duration, in seconds, and TACs for Frontal Cortex (FC), Temporal Cortex (TC), Striatum (STR), Thalamus (THA), Whole Brain (WB), and Cerebellum (CBL)}
  \item{blooddata}{Nested tibble of the blood data. Note that this data is following preprocessing, and does not contain information about automatic and manual sampling. Consists of Time in seconds (ABSS.sec), blood radioactivity following dispersion correction (Cbl.disp.corr) and plasma radioactivity following metabolite correction (Cpl..metabcorr)}
  \item{input}{Nested tibble of the input data, obtained from the blooddata with the blood_interp command}
  \item{Genotype}{Refers to the High Affinity Binder (HAB) or Medium Affinity Binder (MAB) genotype}
  \item{Comment}{Descriptor of irregularities which may have occurred during measurement}
  \item{bodyMass}{Body mass, in kg}
  \item{injRad}{Injected radioactivity}
}}
\source{
\url{https://ejnmmires.springeropen.com/articles/10.1186/s13550-017-0304-1}
}
\usage{
pbr28
}
\description{
A dataset containing the time activity curves (TACs) and blood
data from a study of [11C]PBR28.
}
\references{
Matheson, Granville J, Plav\'en-Sigray, Pontus, Forsberg, Anton, Varrone, Andrea, Farde, Lars & Cervenka, Simon (2017). Assessment of simplified ratio-based approaches for quantification of PET [11 C] PBR28 data. EJNMMI research, 7, 58.
}
\keyword{datasets}
