% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_1tcm.R
\name{plot_1tcmfit}
\alias{plot_1tcmfit}
\title{Plot: One Tissue Compartment Model}
\usage{
plot_1tcmfit(onetcmout, roiname)
}
\arguments{
\item{onetcmout}{The output object of the 1TCM fitting procedure.}

\item{roiname}{Optional. The name of the Target Region to see it on the plot.}
}
\value{
A ggplot2 object of the plot.
}
\description{
Function to visualise the fit of the One Tissue Compartment Model to data.
}
\examples{
plot_1tcmfit(onetcmout)

}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}

