% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_miscfuncs.R
\name{plot_kinfit}
\alias{plot_kinfit}
\title{Plot Kinetic Model Fit: Generic Function}
\usage{
plot_kinfit(modelout, ...)
}
\arguments{
\item{modelout}{The output object of the model fitting procedure.}

\item{...}{Additional optional arguments.}
}
\value{
A ggplot2 object of the plot.
}
\description{
Function to plot the output of a kinetic model. This function calls the specific plotting functions
for each model based on the output of the model.
}
\details{
This function uses the \code{out$model} name to call the correct function to plot the model fit.
}
\examples{
loganout <- Loganplot(t_tac, tac, input, 10, weights)
plot_kinfit(loganout)

srtmout <- srtm(t_tac, reftac, roitac)
plot_kinfit(srtmout, roiname = 'Putamen', refname = 'Cerebellum')

}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}

