% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_LEGA.R
\name{plot_LEGAfit}
\alias{plot_LEGAfit}
\title{Plot: Likelihood Estimation in Graphical Analysis (LEGA)}
\usage{
plot_LEGAfit(LEGAout, roiname = NULL)
}
\arguments{
\item{LEGAout}{The output object of the LEGA fitting procedure.}

\item{roiname}{Optional. The name of the Target Region to see it on the plot.}
}
\value{
A ggplot2 object of the plot.
}
\description{
Function to visualise the fit of the LEGA model to data. Because LEGA models
the original (untransformed) tissue time activity curve, the fit is shown in
the original TAC domain.
}
\examples{
data(pbr28)

t_tac <- pbr28$tacs[[2]]$Times / 60
tac <- pbr28$tacs[[2]]$FC
weights <- pbr28$tacs[[2]]$Weights
dur <- pbr28$tacs[[2]]$Duration/60

input <- blood_interp(
  pbr28$procblood[[2]]$Time / 60, pbr28$procblood[[2]]$Cbl_dispcorr,
  pbr28$procblood[[2]]$Time / 60, pbr28$procblood[[2]]$Cpl_metabcorr,
  t_parentfrac = 1, parentfrac = 1
)

fit <- LEGA(t_tac, dur, tac, input, 10, weights)
plot_LEGAfit(fit)
}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
