% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_ma1.R
\name{plot_ma1fit}
\alias{plot_ma1fit}
\title{Plot: Ichise Multilinear Analysis 1}
\usage{
plot_ma1fit(ma1out, roiname = NULL)
}
\arguments{
\item{ma1out}{The output object of the MA1 fitting procedure.}

\item{roiname}{Optional. The name of the Target Region to see it on the plot.}
}
\value{
A ggplot2 object of the plot.
}
\description{
Function to visualise the fit of the MA1 model to data.
}
\examples{
data(pbr28)

t_tac <- pbr28$tacs[[2]]$Times/60
tac <- pbr28$tacs[[2]]$FC
weights <- pbr28$tacs[[2]]$Weights

input <- blood_interp(
  pbr28$procblood[[2]]$Time/60 , pbr28$procblood[[2]]$Cbl_dispcorr,
  pbr28$procblood[[2]]$Time /60 , pbr28$procblood[[2]]$Cpl_metabcorr,
  t_parentfrac = 1, parentfrac = 1 )

fit <- ma1(t_tac, tac, input, 10, weights)
plot_ma1fit(fit)

}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
