% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_mlloganplot.R
\name{plot_mlLoganfit}
\alias{plot_mlLoganfit}
\title{Plot: Multilinear Logan Plot}
\usage{
plot_mlLoganfit(mlloganout, roiname = NULL)
}
\arguments{
\item{mlloganout}{The output object of the multilinear Logan Plot fitting procedure.}

\item{roiname}{Optional. The name of the Target Region to see it on the plot.}
}
\value{
A ggplot2 object of the plot.
}
\description{
Function to visualise the fit of the multilinear Logan Plot model to data.
}
\examples{
data(pbr28)

t_tac <- pbr28$tacs[[2]]$Times / 60
tac <- pbr28$tacs[[2]]$FC
weights <- pbr28$tacs[[2]]$Weights

input <- blood_interp(
  pbr28$procblood[[2]]$Time / 60, pbr28$procblood[[2]]$Cbl_dispcorr,
  pbr28$procblood[[2]]$Time / 60, pbr28$procblood[[2]]$Cpl_metabcorr,
  t_parentfrac = 1, parentfrac = 1
)

fit <- mlLoganplot(t_tac, tac, input, 10, weights)
plot_mlLoganfit(fit)
}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
