% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_mrtm2.R
\name{plot_mrtm2fit}
\alias{plot_mrtm2fit}
\title{Plot: MRTM2}
\usage{
plot_mrtm2fit(mrtm2out, roiname = NULL, refname = NULL)
}
\arguments{
\item{mrtm2out}{The output object of the mrtm2 fitting procedure.}

\item{roiname}{Optional. The name of the Target Region to see it on the plot.}

\item{refname}{Optional. The name of the Reference Region to see it on the plot.}
}
\value{
A ggplot2 object of the plot.
}
\description{
Function to visualise the fit of the MRTM2 model to data.
}
\examples{

data(pbr28)

t_tac <- pbr28$tacs[[2]]$Times / 60
reftac <- pbr28$tacs[[2]]$CBL
roitac <- pbr28$tacs[[2]]$STR
weights <- pbr28$tacs[[2]]$Weights

fit <- mrtm2(t_tac, reftac, roitac, 0.001, weights = weights)

plot_mrtm2fit(fit)
}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
