% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_patlakplot.R
\name{plot_Patlakfit}
\alias{plot_Patlakfit}
\title{Plot: Patlak Plot}
\usage{
plot_Patlakfit(patlakout, roiname = NULL)
}
\arguments{
\item{patlakout}{The output object of the Patlak Plot fitting procedure.}

\item{roiname}{Optional. The name of the Target Region to see it on the plot.}
}
\value{
A ggplot2 object of the plot.
}
\description{
Function to visualise the fit of the Patlak Plot model to data.
}
\examples{
plot_Patlakfit(patlakout)

}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
