% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_refmllogan.R
\name{plot_refmlLoganfit}
\alias{plot_refmlLoganfit}
\title{Plot: Non-Invasive Multilinear Logan Plot}
\usage{
plot_refmlLoganfit(refmlloganout, roiname = NULL)
}
\arguments{
\item{refmlloganout}{The output object of the refmlLogan fitting procedure.}

\item{roiname}{Optional. The name of the Target Region to see it on the plot.}
}
\value{
A ggplot2 object of the plot.
}
\description{
Function to visualise the fit of the refmlLogan model to data.
}
\examples{
# Note: Reference region models, and irreversible binding models, should not
# be used for PBR28 - this is just to demonstrate function

data(pbr28)

t_tac <- pbr28$tacs[[2]]$Times/60
reftac <- pbr28$tacs[[2]]$CBL
roitac <- pbr28$tacs[[2]]$STR
weights <- pbr28$tacs[[2]]$Weights

fit <- refmlLogan(t_tac, reftac, roitac, k2prime=0.1, tstarIncludedFrames=10, weights=weights)

plot_refmlLoganfit(fit)

}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
