% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_miscfuncs.R
\name{plot_residuals}
\alias{plot_residuals}
\title{Plot Residuals of a Model}
\usage{
plot_residuals(outputobject)
}
\arguments{
\item{outputobject}{The output of a kinetic model, including a fit object \code{outputobject$fit}}
}
\value{
A ggplot2 object of the residual plot
}
\description{
Function to visualise the residuals of a model fit.  This function only works for model outputs for 
which a fit object is available called \code{outputobject$fit}.
}
\examples{
srtmout <- srtm(t_tac, reftac, roitac)
plot_residuals(srtmout)

}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}

