% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_srtm_v.R
\name{plot_srtm_vfit}
\alias{plot_srtm_vfit}
\title{Plot: Simplified Reference Tissue Model with Blood Volumes}
\usage{
plot_srtm_vfit(srtmvout, roiname = NULL, refname = NULL)
}
\arguments{
\item{srtmvout}{The output object of the SRTM fitting procedure.}

\item{roiname}{Optional. The name of the Target Region to see it on the plot.}

\item{refname}{Optional. The name of the Reference Region to see it on the plot.}
}
\value{
A ggplot2 object of the plot.
}
\description{
Function to visualise the fit of the SRTM_V model to data.
}
\examples{
# Note: Reference region models, and irreversible binding models, should not
# be used for PBR28 - this is just to demonstrate function

data(pbr28)

t_tac <- pbr28$tacs[[2]]$Times / 60
reftac <- pbr28$tacs[[2]]$CBL
roitac <- pbr28$tacs[[2]]$STR
weights <- pbr28$tacs[[2]]$Weights

input <- pbr28$input[[2]]

newvals <- shift_timings(
  t_tac,
  roitac,
  input,
  inpshift = 0
)

bloodtac <- pracma::interp1(newvals$input$Time, newvals$input$Blood, t_tac)

fit <- srtm_v(t_tac, reftac, roitac, bloodtac, weights)

plot_srtm_vfit(fit)
}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
