% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_srtm.R
\name{plot_srtmfit}
\alias{plot_srtmfit}
\title{Plot: Simplified Reference Tissue Model}
\usage{
plot_srtmfit(srtmout, roiname = NULL, refname = NULL)
}
\arguments{
\item{srtmout}{The output object of the SRTM fitting procedure.}

\item{roiname}{Optional. The name of the Target Region to see it on the plot.}

\item{refname}{Optional. The name of the Reference Region to see it on the plot.}
}
\value{
A ggplot2 object of the plot.
}
\description{
Function to visualise the fit of the SRTM model to data.
}
\examples{

#' # Note: Reference region models, and irreversible binding models, should not
# be used for PBR28 - this is just to demonstrate function

data(pbr28)

t_tac <- pbr28$tacs[[2]]$Times / 60
reftac <- pbr28$tacs[[2]]$CBL
roitac <- pbr28$tacs[[2]]$STR
weights <- pbr28$tacs[[2]]$Weights

fit <- srtm(t_tac, reftac, roitac)

plot_srtmfit(fit)
}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
