% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_srtm.R
\name{plot_srtmfit}
\alias{plot_srtmfit}
\title{Plot: Simplified Reference Tissue Model}
\usage{
plot_srtmfit(srtmout, roiname = NULL, refname = NULL)
}
\arguments{
\item{srtmout}{The output object of the SRTM fitting procedure.}

\item{roiname}{Optional. The name of the Target Region to see it on the plot.}

\item{refname}{Optional. The name of the Reference Region to see it on the plot.}
}
\value{
A ggplot2 object of the plot.
}
\description{
Function to visualise the fit of the SRTM model to data.
}
\examples{
data(simref)

t_tac <- simref$tacs[[2]]$Times
reftac <- simref$tacs[[2]]$Reference
roitac <- simref$tacs[[2]]$ROI1
weights <- simref$tacs[[2]]$Weights

fit <- srtm(t_tac, reftac, roitac, weights, frameStartEnd = c(1, 33))

plot_srtmfit(fit)
}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
