% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_refPatlak.R
\name{refPatlak}
\alias{refPatlak}
\title{Patlak Reference Tissue Model}
\usage{
refPatlak(t_tac, reftac, roitac, tstarIncludedFrames, weights, frameStartEnd)
}
\arguments{
\item{t_tac}{Numeric vector of times for each frame in minutes. We use the time halfway through the frame as well as a 
zero. If a time zero frame is not included, it will be added.}

\item{reftac}{Numeric vector of radioactivity concentrations in the reference tissue for each frame. We include zero at 
time zero: if not included, it is added.}

\item{roitac}{Numeric vector of radioactivity concentrations in the target tissue for each frame. We include zero at time 
zero: if not included, it is added.}

\item{tstarIncludedFrames}{The number of frames to be used in the regression model, i.e. the number of frames for which 
the function is linear after pseudo-equilibrium is reached. This is a count from the end of the measurement, so a value of 
10 means that last 10 frames will be used. This value can be estimated using \code{refPatlak_tstar}.}

\item{weights}{Optional. Numeric vector of the weights assigned to each frame in the fitting. We include zero at time zero: 
if not included, it is added. If not specified, uniform weights will be used.}

\item{frameStartEnd}{Optional: This allows one to specify the beginning and final frame to use for modelling, e.g. c(1,20). 
This is to assess time stability.}
}
\value{
A list with a data frame of the fitted parameters \code{out$par}, the model fit object \code{out$fit}, a dataframe 
containing the TACs of the data \code{out$tacs}, a dataframe containing the TACs of the fitted values \code{out$fitvals}, 
a vector of the weights \code{out$weights}, and the specified tstarIncludedFrames value \code{out$tstarIncludedFrames}
}
\description{
Function to fit the Patlak Reference Tissue Model of Patlak & Blasbert (1985) to data.
}
\examples{
refPatlak(t_tac, reftac, roitac, tstarIncludedFrames=10, weights=weights)

}
\references{
Patlak CS, Blasberg RG. Graphical evaluation of blood-to-brain transfer constants from multiple-time uptake data. Generalizations. Journal of Cerebral Blood Flow & Metabolism. 1985 Dec 1;5(4):584-90.
}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
