% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_refPatlak.R
\name{refPatlak_tstar}
\alias{refPatlak_tstar}
\title{Tstar Finder: Patlak Reference Tissue Model}
\usage{
refPatlak_tstar(t_tac, reftac, lowroi, medroi, highroi, filename,
  gridbreaks = 2)
}
\arguments{
\item{t_tac}{Numeric vector of times for each frame in minutes. We use the time halfway through the frame as well as a
zero. If a time zero frame is not included, it will be added.}

\item{reftac}{Numeric vector of radioactivity concentrations in the reference tissue for each frame.}

\item{lowroi}{Numeric vector of radioactivity concentrations in a target tissue for each frame. This should be from a ROI with low binding.}

\item{medroi}{Numeric vector of radioactivity concentrations in a target tissue for each frame. This should be from a ROI with medium binding.}

\item{highroi}{Numeric vector of radioactivity concentrations in a target tissue for each frame. This should be from a ROI with high binding.}

\item{filename}{The name of the output image: filename_refPatlak.jpeg}

\item{gridbreaks}{Optional. The size of the grid in the plots. Default: 2.}
}
\value{
Saves a jpeg of the plots as filename_refPatlak.jpeg
}
\description{
Function to identify where t* is for the Patlak Reference Tissue Model.
}
\examples{
\dontrun{
refPatlak_tstar(t_tac, reftac, taclow, tacmed, tachigh, 'demonstration')
}

}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
