% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_SUV.R
\name{suv}
\alias{suv}
\title{Standardised Uptake Value}
\usage{
suv(
  tac,
  t_tac = NULL,
  dur = NULL,
  injRad = 1,
  bodymass = 1,
  frameStartEnd = NULL,
  timeStartEnd = NULL
)
}
\arguments{
\item{tac}{Numeric vector of radioactivity concentrations in the target tissue for each frame. We include zero at time
zero: if not included, it is added.}

\item{t_tac}{Numeric vector of times for each frame in minutes. We use the time halfway through the frame as well as a
zero. If a time zero frame is not included, it will be added.}

\item{dur}{Numeric vector of durations for each frame in minutes. This will be used instead of the middle points of the frames if provided.}

\item{injRad}{The injected radioactivity.  If not included, this will be set to 1 in case one is using SUV ratios.}

\item{bodymass}{The body mass of the participant. If not included, this will be set to 1 in case one is using SUV ratios.}

\item{frameStartEnd}{Optional: This allows one to specify the beginning and final frame to use for modelling, e.g. c(1,20).
This can be used to assess time stability for example.}

\item{timeStartEnd}{Optional. This allows one to specify the beginning and end time point instead of defining the frame numbers using frameStartEnd. This function will restrict the model to all time frames whose t_tac is between the values, i.e. c(0,5) will select all frames with midtimes during the first 5 minutes. Note that this requires t_tac.}
}
\value{
A list with a data frame of the calculated parameters \code{out$par} and a dataframe containing the TACs both of the
original data and the transformed values \code{out$tacs}. \code{out$tacs} contains the whole measured TAC, with a logical
\code{Included} column marking the frames which the integral was taken over, i.e. those selected by
\code{frameStartEnd} or \code{timeStartEnd}.

\code{out$par} contains \code{SUV}, the mean SUV over the frames which were included; \code{SUV_AUC}, its integral over
those frames; \code{SUV_denominator}, the injected radioactivity divided by the body mass which was applied; and
\code{window_duration} and \code{n_frames} describing what was integrated. With the default \code{injRad} and
\code{bodymass} of 1, the denominator is 1 and the outcomes are in units of radioactivity concentration.
}
\description{
Function to calculate the standardised uptake value and its integral.
}
\examples{
data(pbr28)

t_tac <- pbr28$tacs[[2]]$Times / 60
dur <- pbr28$tacs[[2]]$Duration / 60
tac <- pbr28$tacs[[2]]$FC


fit1 <- suv(tac, t_tac, injRad = 150, bodymass = 85)
fit2 <- suv(tac, dur = dur, injRad = 150, bodymass = 85)
fit3 <- suv(tac, t_tac = t_tac, dur = dur, injRad = 150, bodymass = 85)
fit4 <- suv(tac, t_tac = t_tac, dur = dur, injRad = 150, bodymass = 85,
   frameStartEnd = c(30,33))
fit5 <- suv(tac, t_tac = t_tac, dur = dur, injRad = 150, bodymass = 85,
   timeStartEnd = c(40,61))
}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
