% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_SUV.R
\name{SUV}
\alias{SUV}
\title{Standardised Uptake Value}
\usage{
SUV(t_tac, tac, injRad = 1, bodymass = 1, frameStartEnd)
}
\arguments{
\item{t_tac}{Numeric vector of times for each frame in minutes. We use the time halfway through the frame as well as a 
zero. If a time zero frame is not included, it will be added.}

\item{tac}{Numeric vector of radioactivity concentrations in the target tissue for each frame. We include zero at time 
zero: if not included, it is added.}

\item{injRad}{The injected radioactivity.  If not included, this will be set to 1 in case one is using SUV ratios.}

\item{bodymass}{The body mass of the participant. If not included, this will be set to 1 in case one is using SUV ratios.}

\item{frameStartEnd}{Optional: This allows one to specify the beginning and final frame to use for modelling, e.g. c(1,20). 
This is to assess time stability.}
}
\value{
A list with a data frame of the calculated parameters \code{out$par} and a dataframe containing the TACs both of the 
original data and the transformed values \code{out$tacs}
}
\description{
Function to calculate the standardised uptake value and its integral.
}
\examples{
SUV(t_tac, tac, injRad=150, bodymass=85)

}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
