% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitr_weights.R
\name{weights_create}
\alias{weights_create}
\title{Create weights using durations and counts}
\usage{
weights_create(
  t_start,
  t_end,
  tac,
  radioisotope = c("C11", "O15", "F18"),
  method = 2,
  minweight = 0.7,
  weight_checkn = 5
)
}
\arguments{
\item{t_start}{The starting times of the frames in minutes.}

\item{t_end}{The end times of the frames in minutes.}

\item{tac}{The time activity curve.}

\item{radioisotope}{The radioisotope.}

\item{method}{Which method should be used? 1 represents duration^2 /
(tac_uncorrected). 2 represents sqrt(durations*tac_uncorrected). 3
represents duration / tac. 4 represents sqrt(durations). 5 represents
durations * exp((-ln(2)) / halflife ). 6 represents durations /
tac. 7 represents durations.}

\item{minweight}{The minimum weight. Weights will be calculated as a fraction
between this value and 1. A zero frame with duration=0 will be set to 0
though.}

\item{weight_checkn}{The number of values of the weights to check to make
sure that things haven't gone terribly wrong. It will check that this
number of weights are all above half of the maximum weight.}
}
\value{
The vector of weights.
}
\description{
This set of methods allows for creating a weighting scheme based on TAC data
alone. It is advised to calculate weights once for each PET measurement,
using a large region (e.g. whole brain).
}
\examples{
data(pbr28)
s1 <- pbr28$tacs[[1]]

# Assuming the data were not decay-corrected (they are)
weights_create(
  s1$StartTime/60,
  (s1$StartTime + s1$Duration)/60,
  radioisotope = "C11",
  tac = s1$WB)
}
